% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bswap_rmv_liquidity_preview_single.R
\docType{class}
\name{BswapRmvLiquidityPreviewSingle}
\alias{BswapRmvLiquidityPreviewSingle}
\title{BswapRmvLiquidityPreviewSingle}
\format{
An \code{R6Class} generator object
}
\description{
BswapRmvLiquidityPreviewSingle Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{quoteAsset}}{character}

\item{\code{quoteAmt}}{integer}

\item{\code{price}}{numeric}

\item{\code{slippage}}{numeric}

\item{\code{fee}}{numeric}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BswapRmvLiquidityPreviewSingle$new()}}
\item \href{#method-toJSON}{\code{BswapRmvLiquidityPreviewSingle$toJSON()}}
\item \href{#method-fromJSON}{\code{BswapRmvLiquidityPreviewSingle$fromJSON()}}
\item \href{#method-toJSONString}{\code{BswapRmvLiquidityPreviewSingle$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BswapRmvLiquidityPreviewSingle$fromJSONString()}}
\item \href{#method-clone}{\code{BswapRmvLiquidityPreviewSingle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BswapRmvLiquidityPreviewSingle$new(
  quoteAsset,
  quoteAmt,
  price,
  slippage,
  fee,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BswapRmvLiquidityPreviewSingle$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BswapRmvLiquidityPreviewSingle$fromJSON(BswapRmvLiquidityPreviewSingleJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BswapRmvLiquidityPreviewSingle$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BswapRmvLiquidityPreviewSingle$fromJSONString(
  BswapRmvLiquidityPreviewSingleJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BswapRmvLiquidityPreviewSingle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
