% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2_c_api.R
\docType{class}
\name{C2CApi}
\alias{C2CApi}
\title{C2C operations}
\format{
An \code{R6Class} generator object
}
\description{
binanceRapi.C2C
}
\section{Methods}{

\describe{
\strong{ SapiV1C2cOrderMatchListUserOrderHistoryGet } \emph{ Get C2C Trade History (USER_DATA) }
- If startTimestamp and endTimestamp are not sent, the recent 30-day data will be returned. - The max interval between startTimestamp and endTimestamp is 30 days.  Weight(IP): 1

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } trade.type Enum < [BUY, SELL] >
\item \emph{ @param } start.timestamp integer
\item \emph{ @param } end.timestamp integer
\item \emph{ @param } page integer
\item \emph{ @param } rows integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse200111} \cr


\item status code : 200 | Trades history

\item return type : InlineResponse200111 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  SapiV1C2cOrderMatchListUserOrderHistoryGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.trade.type <- 'trade.type_example' # character | 
var.start.timestamp <- 56 # integer | UTC timestamp in ms
var.end.timestamp <- 56 # integer | UTC timestamp in ms
var.page <- 1 # integer | Default 1
var.rows <- 56 # integer | default 100, max 100
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get C2C Trade History (USER_DATA)
api.instance <- C2CApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1C2cOrderMatchListUserOrderHistoryGet(var.timestamp, var.signature, trade.type=var.trade.type, start.timestamp=var.start.timestamp, end.timestamp=var.end.timestamp, page=var.page, rows=var.rows, recv.window=var.recv.window)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{C2CApi$new()}}
\item \href{#method-SapiV1C2cOrderMatchListUserOrderHistoryGet}{\code{C2CApi$SapiV1C2cOrderMatchListUserOrderHistoryGet()}}
\item \href{#method-SapiV1C2cOrderMatchListUserOrderHistoryGetWithHttpInfo}{\code{C2CApi$SapiV1C2cOrderMatchListUserOrderHistoryGetWithHttpInfo()}}
\item \href{#method-clone}{\code{C2CApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C2CApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1C2cOrderMatchListUserOrderHistoryGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1C2cOrderMatchListUserOrderHistoryGet}{}}}
\subsection{Method \code{SapiV1C2cOrderMatchListUserOrderHistoryGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C2CApi$SapiV1C2cOrderMatchListUserOrderHistoryGet(
  timestamp,
  signature,
  trade.type = NULL,
  start.timestamp = NULL,
  end.timestamp = NULL,
  page = NULL,
  rows = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1C2cOrderMatchListUserOrderHistoryGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1C2cOrderMatchListUserOrderHistoryGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1C2cOrderMatchListUserOrderHistoryGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C2CApi$SapiV1C2cOrderMatchListUserOrderHistoryGetWithHttpInfo(
  timestamp,
  signature,
  trade.type = NULL,
  start.timestamp = NULL,
  end.timestamp = NULL,
  page = NULL,
  rows = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C2CApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
