% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_api.R
\docType{class}
\name{ConvertApi}
\alias{ConvertApi}
\title{Convert operations}
\format{
An \code{R6Class} generator object
}
\description{
binanceRapi.Convert
}
\section{Methods}{

\describe{
\strong{ SapiV1ConvertTradeFlowGet } \emph{ Get Convert Trade History (USER_DATA) }
- The max interval between startTime and endTime is 30 days.  Weight(UID): 3000

\itemize{
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } limit integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse200121} \cr


\item status code : 200 | Convert Trade History

\item return type : InlineResponse200121 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  SapiV1ConvertTradeFlowGet  ####################

library(binanceRapi)
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.limit <- 100 # integer | default 100, max 1000
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get Convert Trade History (USER_DATA)
api.instance <- ConvertApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1ConvertTradeFlowGet(var.start.time, var.end.time, var.timestamp, var.signature, limit=var.limit, recv.window=var.recv.window)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConvertApi$new()}}
\item \href{#method-SapiV1ConvertTradeFlowGet}{\code{ConvertApi$SapiV1ConvertTradeFlowGet()}}
\item \href{#method-SapiV1ConvertTradeFlowGetWithHttpInfo}{\code{ConvertApi$SapiV1ConvertTradeFlowGetWithHttpInfo()}}
\item \href{#method-clone}{\code{ConvertApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConvertApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1ConvertTradeFlowGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1ConvertTradeFlowGet}{}}}
\subsection{Method \code{SapiV1ConvertTradeFlowGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConvertApi$SapiV1ConvertTradeFlowGet(
  start.time,
  end.time,
  timestamp,
  signature,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1ConvertTradeFlowGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1ConvertTradeFlowGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1ConvertTradeFlowGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConvertApi$SapiV1ConvertTradeFlowGetWithHttpInfo(
  start.time,
  end.time,
  timestamp,
  signature,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConvertApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
