% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{Credentials}
\alias{Credentials}
\title{R6 class for Binance credentials}
\description{
R6 class for Binance credentials

R6 class for Binance credentials
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Credentials$new()}}
\item \href{#method-sign}{\code{Credentials$sign()}}
\item \href{#method-setSecret}{\code{Credentials$setSecret()}}
\item \href{#method-setKey}{\code{Credentials$setKey()}}
\item \href{#method-checkCredentials}{\code{Credentials$checkCredentials()}}
\item \href{#method-clone}{\code{Credentials$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credentials$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sign"></a>}}
\if{latex}{\out{\hypertarget{method-sign}{}}}
\subsection{Method \code{sign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credentials$sign(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSecret"></a>}}
\if{latex}{\out{\hypertarget{method-setSecret}{}}}
\subsection{Method \code{setSecret()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credentials$setSecret(secret)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setKey"></a>}}
\if{latex}{\out{\hypertarget{method-setKey}{}}}
\subsection{Method \code{setKey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credentials$setKey(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkCredentials"></a>}}
\if{latex}{\out{\hypertarget{method-checkCredentials}{}}}
\subsection{Method \code{checkCredentials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credentials$checkCredentials()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credentials$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
