% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiat_api.R
\docType{class}
\name{FiatApi}
\alias{FiatApi}
\title{Fiat operations}
\format{
An \code{R6Class} generator object
}
\description{
binanceRapi.Fiat
}
\section{}{
NA
}

\examples{
\dontrun{
####################  SapiV1FiatOrdersGet  ####################

library(binanceRapi)
var.transaction.type <- '0' # character | 0-deposit, 1-withdraw
var.begin.time <- 1626144956000 # integer | 
var.end.time <- 56 # integer | UTC timestamp in ms
var.page <- 1 # integer | Default 1
var.rows <- 300 # integer | Default 100, max 500
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Fiat Deposit/Withdraw History (USER_DATA)
api.instance <- FiatApi$new()

result <- api.instance$SapiV1FiatOrdersGet(var.transaction.type, var.timestamp, var.signature, begin.time=var.begin.time, end.time=var.end.time, page=var.page, rows=var.rows, recv.window=var.recv.window)


####################  SapiV1FiatPaymentsGet  ####################

library(binanceRapi)
var.transaction.type <- '0' # character | 0-deposit, 1-withdraw
var.begin.time <- 1626144956000 # integer | 
var.end.time <- 56 # integer | UTC timestamp in ms
var.page <- 1 # integer | Default 1
var.rows <- 300 # integer | Default 100, max 500
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Fiat Payments History (USER_DATA)
api.instance <- FiatApi$new()

result <- api.instance$SapiV1FiatPaymentsGet(var.transaction.type, var.timestamp, var.signature, begin.time=var.begin.time, end.time=var.end.time, page=var.page, rows=var.rows, recv.window=var.recv.window)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FiatApi$new()}}
\item \href{#method-SapiV1FiatOrdersGet}{\code{FiatApi$SapiV1FiatOrdersGet()}}
\item \href{#method-SapiV1FiatOrdersGetWithHttpInfo}{\code{FiatApi$SapiV1FiatOrdersGetWithHttpInfo()}}
\item \href{#method-SapiV1FiatPaymentsGet}{\code{FiatApi$SapiV1FiatPaymentsGet()}}
\item \href{#method-SapiV1FiatPaymentsGetWithHttpInfo}{\code{FiatApi$SapiV1FiatPaymentsGetWithHttpInfo()}}
\item \href{#method-clone}{\code{FiatApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FiatApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1FiatOrdersGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1FiatOrdersGet}{}}}
\subsection{Method \code{SapiV1FiatOrdersGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FiatApi$SapiV1FiatOrdersGet(
  transaction.type,
  begin.time = NULL,
  end.time = NULL,
  page = NULL,
  rows = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1FiatOrdersGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1FiatOrdersGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1FiatOrdersGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FiatApi$SapiV1FiatOrdersGetWithHttpInfo(
  transaction.type,
  begin.time = NULL,
  end.time = NULL,
  page = NULL,
  rows = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1FiatPaymentsGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1FiatPaymentsGet}{}}}
\subsection{Method \code{SapiV1FiatPaymentsGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FiatApi$SapiV1FiatPaymentsGet(
  transaction.type,
  begin.time = NULL,
  end.time = NULL,
  page = NULL,
  rows = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1FiatPaymentsGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1FiatPaymentsGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1FiatPaymentsGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FiatApi$SapiV1FiatPaymentsGetWithHttpInfo(
  transaction.type,
  begin.time = NULL,
  end.time = NULL,
  page = NULL,
  rows = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FiatApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
