% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response200102_data.R
\docType{class}
\name{InlineResponse200102Data}
\alias{InlineResponse200102Data}
\title{InlineResponse200102Data}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse200102Data Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fifteenMinHashRate}}{character}

\item{\code{dayHashRate}}{character}

\item{\code{validNum}}{integer}

\item{\code{invalidNum}}{integer}

\item{\code{profitToday}}{\link{InlineResponse200102DataProfitToday}}

\item{\code{profitYesterday}}{\link{InlineResponse200102DataProfitToday}}

\item{\code{userName}}{character}

\item{\code{unit}}{character}

\item{\code{algo}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse200102Data$new()}}
\item \href{#method-toJSON}{\code{InlineResponse200102Data$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse200102Data$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse200102Data$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse200102Data$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse200102Data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200102Data$new(
  fifteenMinHashRate,
  dayHashRate,
  validNum,
  invalidNum,
  profitToday,
  profitYesterday,
  userName,
  unit,
  algo,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200102Data$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200102Data$fromJSON(InlineResponse200102DataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200102Data$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200102Data$fromJSONString(InlineResponse200102DataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200102Data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
