% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response200115.R
\docType{class}
\name{InlineResponse200115}
\alias{InlineResponse200115}
\title{InlineResponse200115}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse200115 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{swapId}}{integer}

\item{\code{swapTime}}{integer}

\item{\code{status}}{integer}

\item{\code{quoteAsset}}{character}

\item{\code{baseAsset}}{character}

\item{\code{quoteQty}}{numeric}

\item{\code{baseQty}}{numeric}

\item{\code{price}}{numeric}

\item{\code{fee}}{numeric}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse200115$new()}}
\item \href{#method-toJSON}{\code{InlineResponse200115$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse200115$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse200115$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse200115$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse200115$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200115$new(
  swapId,
  swapTime,
  status,
  quoteAsset,
  baseAsset,
  quoteQty,
  baseQty,
  price,
  fee,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200115$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200115$fromJSON(InlineResponse200115Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200115$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200115$fromJSONString(InlineResponse200115Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200115$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
