% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response200117.R
\docType{class}
\name{InlineResponse200117}
\alias{InlineResponse200117}
\title{InlineResponse200117}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse200117 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{poolId}}{integer}

\item{\code{poolNmae}}{character}

\item{\code{updateTime}}{integer}

\item{\code{liquidity}}{\link{SapiV1BswapPoolConfigureLiquidity}}

\item{\code{assetConfigure}}{\link{SapiV1BswapPoolConfigureAssetConfigure}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse200117$new()}}
\item \href{#method-toJSON}{\code{InlineResponse200117$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse200117$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse200117$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse200117$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse200117$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200117$new(
  poolId,
  poolNmae,
  updateTime,
  liquidity,
  assetConfigure,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200117$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200117$fromJSON(InlineResponse200117Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200117$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200117$fromJSONString(InlineResponse200117Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200117$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
