% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response200118_data.R
\docType{class}
\name{InlineResponse200118Data}
\alias{InlineResponse200118Data}
\title{InlineResponse200118Data}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse200118Data Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{orderNumber}}{character}

\item{\code{advNo}}{character}

\item{\code{tradeType}}{character}

\item{\code{asset}}{character}

\item{\code{fiat}}{character}

\item{\code{fiatSymbol}}{character}

\item{\code{amount}}{character}

\item{\code{totalPrice}}{character}

\item{\code{unitPrice}}{character}

\item{\code{orderStatus}}{character}

\item{\code{createTime}}{integer}

\item{\code{commission}}{character}

\item{\code{counterPartNickName}}{character}

\item{\code{advertisementRole}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse200118Data$new()}}
\item \href{#method-toJSON}{\code{InlineResponse200118Data$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse200118Data$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse200118Data$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse200118Data$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse200118Data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200118Data$new(
  orderNumber,
  advNo,
  tradeType,
  asset,
  fiat,
  fiatSymbol,
  amount,
  totalPrice,
  unitPrice,
  orderStatus,
  createTime,
  commission,
  counterPartNickName,
  advertisementRole,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200118Data$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200118Data$fromJSON(InlineResponse200118DataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200118Data$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200118Data$fromJSONString(InlineResponse200118DataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200118Data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
