% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response200121_list.R
\docType{class}
\name{InlineResponse200121List}
\alias{InlineResponse200121List}
\title{InlineResponse200121List}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse200121List Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{quoteId}}{character}

\item{\code{orderId}}{integer}

\item{\code{orderStatus}}{character}

\item{\code{fromAsset}}{character}

\item{\code{fromAmount}}{character}

\item{\code{toAsset}}{character}

\item{\code{toAmount}}{character}

\item{\code{ratio}}{character}

\item{\code{inverseRatio}}{character}

\item{\code{createTime}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse200121List$new()}}
\item \href{#method-toJSON}{\code{InlineResponse200121List$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse200121List$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse200121List$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse200121List$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse200121List$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200121List$new(
  quoteId,
  orderId,
  orderStatus,
  fromAsset,
  fromAmount,
  toAsset,
  toAmount,
  ratio,
  inverseRatio,
  createTime,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200121List$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200121List$fromJSON(InlineResponse200121ListJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200121List$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200121List$fromJSONString(InlineResponse200121ListJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200121List$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
