% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response200125_list.R
\docType{class}
\name{InlineResponse200125List}
\alias{InlineResponse200125List}
\title{InlineResponse200125List}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse200125List Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{network}}{character}

\item{\code{txID}}{character}

\item{\code{contractAdrress}}{character}

\item{\code{tokenId}}{character}

\item{\code{timestamp}}{integer}

\item{\code{fee}}{numeric}

\item{\code{feeAsset}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse200125List$new()}}
\item \href{#method-toJSON}{\code{InlineResponse200125List$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse200125List$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse200125List$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse200125List$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse200125List$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200125List$new(
  network,
  txID,
  contractAdrress,
  tokenId,
  timestamp,
  fee,
  feeAsset,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200125List$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200125List$fromJSON(InlineResponse200125ListJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200125List$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200125List$fromJSONString(InlineResponse200125ListJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200125List$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
