% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20017_rows.R
\docType{class}
\name{InlineResponse20017Rows}
\alias{InlineResponse20017Rows}
\title{InlineResponse20017Rows}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20017Rows Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isolatedSymbol}}{character}

\item{\code{asset}}{character}

\item{\code{interest}}{character}

\item{\code{interestAccuredTime}}{integer}

\item{\code{interestRate}}{character}

\item{\code{principal}}{character}

\item{\code{type}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20017Rows$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20017Rows$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20017Rows$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20017Rows$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20017Rows$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20017Rows$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20017Rows$new(
  isolatedSymbol,
  asset,
  interest,
  interestAccuredTime,
  interestRate,
  principal,
  type,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20017Rows$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20017Rows$fromJSON(InlineResponse20017RowsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20017Rows$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20017Rows$fromJSONString(InlineResponse20017RowsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20017Rows$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
