% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20018_rows.R
\docType{class}
\name{InlineResponse20018Rows}
\alias{InlineResponse20018Rows}
\title{InlineResponse20018Rows}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20018Rows Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{avgPrice}}{character}

\item{\code{executedQty}}{character}

\item{\code{orderId}}{integer}

\item{\code{price}}{character}

\item{\code{qty}}{character}

\item{\code{side}}{character}

\item{\code{symbol}}{character}

\item{\code{timeInForce}}{character}

\item{\code{isIsolated}}{character}

\item{\code{updatedTime}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20018Rows$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20018Rows$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20018Rows$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20018Rows$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20018Rows$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20018Rows$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20018Rows$new(
  avgPrice,
  executedQty,
  orderId,
  price,
  qty,
  side,
  symbol,
  timeInForce,
  isIsolated,
  updatedTime,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20018Rows$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20018Rows$fromJSON(InlineResponse20018RowsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20018Rows$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20018Rows$fromJSONString(InlineResponse20018RowsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20018Rows$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
