% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20019_user_assets.R
\docType{class}
\name{InlineResponse20019UserAssets}
\alias{InlineResponse20019UserAssets}
\title{InlineResponse20019UserAssets}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20019UserAssets Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{asset}}{character}

\item{\code{borrowed}}{character}

\item{\code{free}}{character}

\item{\code{interest}}{character}

\item{\code{locked}}{character}

\item{\code{netAsset}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20019UserAssets$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20019UserAssets$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20019UserAssets$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20019UserAssets$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20019UserAssets$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20019UserAssets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20019UserAssets$new(
  asset,
  borrowed,
  free,
  interest,
  locked,
  netAsset,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20019UserAssets$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20019UserAssets$fromJSON(InlineResponse20019UserAssetsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20019UserAssets$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20019UserAssets$fromJSONString(InlineResponse20019UserAssetsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20019UserAssets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
