% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response2001_rate_limits.R
\docType{class}
\name{InlineResponse2001RateLimits}
\alias{InlineResponse2001RateLimits}
\title{InlineResponse2001RateLimits}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse2001RateLimits Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rateLimitType}}{character}

\item{\code{interval}}{character}

\item{\code{intervalNum}}{integer}

\item{\code{limit}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse2001RateLimits$new()}}
\item \href{#method-toJSON}{\code{InlineResponse2001RateLimits$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse2001RateLimits$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse2001RateLimits$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse2001RateLimits$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse2001RateLimits$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001RateLimits$new(
  rateLimitType,
  interval,
  intervalNum,
  limit,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001RateLimits$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001RateLimits$fromJSON(InlineResponse2001RateLimitsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001RateLimits$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001RateLimits$fromJSONString(InlineResponse2001RateLimitsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001RateLimits$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
