% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response2001_symbols.R
\docType{class}
\name{InlineResponse2001Symbols}
\alias{InlineResponse2001Symbols}
\title{InlineResponse2001Symbols}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse2001Symbols Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{symbol}}{character}

\item{\code{status}}{character}

\item{\code{baseAsset}}{character}

\item{\code{baseAssetPrecision}}{integer}

\item{\code{quoteAsset}}{character}

\item{\code{quoteAssetPrecision}}{integer}

\item{\code{baseCommissionPrecision}}{integer}

\item{\code{quoteCommissionPrecision}}{integer}

\item{\code{orderTypes}}{list( character )}

\item{\code{icebergAllowed}}{character}

\item{\code{ocoAllowed}}{character}

\item{\code{quoteOrderQtyMarketAllowed}}{character}

\item{\code{isSpotTradingAllowed}}{character}

\item{\code{isMarginTradingAllowed}}{character}

\item{\code{filters}}{list( \link{InlineResponse2001Filters} )}

\item{\code{permissions}}{list( character )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse2001Symbols$new()}}
\item \href{#method-toJSON}{\code{InlineResponse2001Symbols$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse2001Symbols$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse2001Symbols$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse2001Symbols$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse2001Symbols$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001Symbols$new(
  symbol,
  status,
  baseAsset,
  baseAssetPrecision,
  quoteAsset,
  quoteAssetPrecision,
  baseCommissionPrecision,
  quoteCommissionPrecision,
  orderTypes,
  icebergAllowed,
  ocoAllowed,
  quoteOrderQtyMarketAllowed,
  isSpotTradingAllowed,
  isMarginTradingAllowed,
  filters,
  permissions,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001Symbols$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001Symbols$fromJSON(InlineResponse2001SymbolsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001Symbols$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001Symbols$fromJSONString(InlineResponse2001SymbolsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001Symbols$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
