% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20021.R
\docType{class}
\name{InlineResponse20021}
\alias{InlineResponse20021}
\title{InlineResponse20021}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20021 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{orderListId}}{integer}

\item{\code{contingencyType}}{character}

\item{\code{listStatusType}}{character}

\item{\code{listOrderStatus}}{character}

\item{\code{listClientOrderId}}{character}

\item{\code{transactionTime}}{integer}

\item{\code{symbol}}{character}

\item{\code{isIsolated}}{character}

\item{\code{orders}}{list( \link{InlineResponse2005Orders} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20021$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20021$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20021$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20021$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20021$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20021$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20021$new(
  orderListId,
  contingencyType,
  listStatusType,
  listOrderStatus,
  listClientOrderId,
  transactionTime,
  symbol,
  isIsolated,
  orders,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20021$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20021$fromJSON(InlineResponse20021Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20021$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20021$fromJSONString(InlineResponse20021Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20021$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
