% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20032.R
\docType{class}
\name{InlineResponse20032}
\alias{InlineResponse20032}
\title{InlineResponse20032}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20032 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{amount}}{character}

\item{\code{coin}}{character}

\item{\code{network}}{character}

\item{\code{status}}{integer}

\item{\code{address}}{character}

\item{\code{addressTag}}{character}

\item{\code{txId}}{character}

\item{\code{insertTime}}{integer}

\item{\code{transferType}}{integer}

\item{\code{unlockConfirm}}{character}

\item{\code{confirmTimes}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20032$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20032$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20032$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20032$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20032$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20032$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20032$new(
  amount,
  coin,
  network,
  status,
  address,
  addressTag,
  txId,
  insertTime,
  transferType,
  unlockConfirm,
  confirmTimes,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20032$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20032$fromJSON(InlineResponse20032Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20032$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20032$fromJSONString(InlineResponse20032Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20032$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
