% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20036.R
\docType{class}
\name{InlineResponse20036}
\alias{InlineResponse20036}
\title{InlineResponse20036}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20036 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{address}}{character}

\item{\code{amount}}{character}

\item{\code{applyTime}}{character}

\item{\code{coin}}{character}

\item{\code{id}}{character}

\item{\code{withdrawOrderId}}{character}

\item{\code{network}}{character}

\item{\code{transferType}}{integer}

\item{\code{status}}{integer}

\item{\code{transactionFee}}{character}

\item{\code{confirmNo}}{integer}

\item{\code{info}}{character}

\item{\code{txId}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20036$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20036$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20036$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20036$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20036$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20036$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036$new(
  address,
  amount,
  applyTime,
  coin,
  id,
  withdrawOrderId,
  network,
  transferType,
  status,
  transactionFee,
  confirmNo,
  info,
  txId,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036$fromJSON(InlineResponse20036Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036$fromJSONString(InlineResponse20036Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
