% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20036_data_indicators_btcusdt.R
\docType{class}
\name{InlineResponse20036DataIndicatorsBTCUSDT}
\alias{InlineResponse20036DataIndicatorsBTCUSDT}
\title{InlineResponse20036DataIndicatorsBTCUSDT}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20036DataIndicatorsBTCUSDT Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{i}}{character}

\item{\code{c}}{integer}

\item{\code{v}}{numeric}

\item{\code{t}}{numeric}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20036DataIndicatorsBTCUSDT$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20036DataIndicatorsBTCUSDT$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20036DataIndicatorsBTCUSDT$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20036DataIndicatorsBTCUSDT$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20036DataIndicatorsBTCUSDT$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20036DataIndicatorsBTCUSDT$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036DataIndicatorsBTCUSDT$new(i, c, v, t, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036DataIndicatorsBTCUSDT$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036DataIndicatorsBTCUSDT$fromJSON(
  InlineResponse20036DataIndicatorsBTCUSDTJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036DataIndicatorsBTCUSDT$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036DataIndicatorsBTCUSDT$fromJSONString(
  InlineResponse20036DataIndicatorsBTCUSDTJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20036DataIndicatorsBTCUSDT$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
