% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inline_response20037_user_asset_dribblet_details.R
\docType{class}
\name{InlineResponse20037UserAssetDribbletDetails}
\alias{InlineResponse20037UserAssetDribbletDetails}
\title{InlineResponse20037UserAssetDribbletDetails}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20037UserAssetDribbletDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{transId}}{integer}

\item{\code{serviceChargeAmount}}{character}

\item{\code{amount}}{character}

\item{\code{operateTime}}{integer}

\item{\code{transferedAmount}}{character}

\item{\code{fromAsset}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20037UserAssetDribbletDetails$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20037UserAssetDribbletDetails$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20037UserAssetDribbletDetails$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20037UserAssetDribbletDetails$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20037UserAssetDribbletDetails$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20037UserAssetDribbletDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20037UserAssetDribbletDetails$new(
  transId,
  serviceChargeAmount,
  amount,
  operateTime,
  transferedAmount,
  fromAsset,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20037UserAssetDribbletDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20037UserAssetDribbletDetails$fromJSON(
  InlineResponse20037UserAssetDribbletDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20037UserAssetDribbletDetails$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20037UserAssetDribbletDetails$fromJSONString(
  InlineResponse20037UserAssetDribbletDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20037UserAssetDribbletDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
