% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20038_transfer_result.R
\docType{class}
\name{InlineResponse20038TransferResult}
\alias{InlineResponse20038TransferResult}
\title{InlineResponse20038TransferResult}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20038TransferResult Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{amount}}{character}

\item{\code{fromAsset}}{character}

\item{\code{operateTime}}{integer}

\item{\code{serviceChargeAmount}}{character}

\item{\code{tranId}}{integer}

\item{\code{transferedAmount}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20038TransferResult$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20038TransferResult$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20038TransferResult$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20038TransferResult$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20038TransferResult$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20038TransferResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20038TransferResult$new(
  amount,
  fromAsset,
  operateTime,
  serviceChargeAmount,
  tranId,
  transferedAmount,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20038TransferResult$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20038TransferResult$fromJSON(
  InlineResponse20038TransferResultJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20038TransferResult$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20038TransferResult$fromJSONString(
  InlineResponse20038TransferResultJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20038TransferResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
