% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20039_data_trigger_condition.R
\docType{class}
\name{InlineResponse20039DataTriggerCondition}
\alias{InlineResponse20039DataTriggerCondition}
\title{InlineResponse20039DataTriggerCondition}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20039DataTriggerCondition Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{GCR}}{integer}

\item{\code{IFER}}{integer}

\item{\code{UFR}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20039DataTriggerCondition$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20039DataTriggerCondition$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20039DataTriggerCondition$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20039DataTriggerCondition$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20039DataTriggerCondition$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20039DataTriggerCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20039DataTriggerCondition$new(GCR, IFER, UFR, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20039DataTriggerCondition$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20039DataTriggerCondition$fromJSON(
  InlineResponse20039DataTriggerConditionJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20039DataTriggerCondition$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20039DataTriggerCondition$fromJSONString(
  InlineResponse20039DataTriggerConditionJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20039DataTriggerCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
