% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20040_ctr.R
\docType{class}
\name{InlineResponse20040CTR}
\alias{InlineResponse20040CTR}
\title{InlineResponse20040CTR}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20040CTR Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{minWithdrawAmount}}{character}

\item{\code{depositStatus}}{character}

\item{\code{withdrawFee}}{integer}

\item{\code{withdrawStatus}}{character}

\item{\code{depositTip}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20040CTR$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20040CTR$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20040CTR$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20040CTR$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20040CTR$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20040CTR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20040CTR$new(
  minWithdrawAmount,
  depositStatus,
  withdrawFee,
  withdrawStatus,
  depositTip,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20040CTR$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20040CTR$fromJSON(InlineResponse20040CTRJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20040CTR$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20040CTR$fromJSONString(InlineResponse20040CTRJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20040CTR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
