% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inline_response20055_spot_sub_user_asset_btc_vo_list.R
\docType{class}
\name{InlineResponse20055SpotSubUserAssetBtcVoList}
\alias{InlineResponse20055SpotSubUserAssetBtcVoList}
\title{InlineResponse20055SpotSubUserAssetBtcVoList}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20055SpotSubUserAssetBtcVoList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{character}

\item{\code{totalAsset}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20055SpotSubUserAssetBtcVoList$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20055SpotSubUserAssetBtcVoList$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20055SpotSubUserAssetBtcVoList$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20055SpotSubUserAssetBtcVoList$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20055SpotSubUserAssetBtcVoList$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20055SpotSubUserAssetBtcVoList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20055SpotSubUserAssetBtcVoList$new(email, totalAsset, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20055SpotSubUserAssetBtcVoList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20055SpotSubUserAssetBtcVoList$fromJSON(
  InlineResponse20055SpotSubUserAssetBtcVoListJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20055SpotSubUserAssetBtcVoList$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20055SpotSubUserAssetBtcVoList$fromJSONString(
  InlineResponse20055SpotSubUserAssetBtcVoListJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20055SpotSubUserAssetBtcVoList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
