% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20057.R
\docType{class}
\name{InlineResponse20057}
\alias{InlineResponse20057}
\title{InlineResponse20057}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20057 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{character}

\item{\code{marginLevel}}{character}

\item{\code{totalAssetOfBtc}}{character}

\item{\code{totalLiabilityOfBtc}}{character}

\item{\code{totalNetAssetOfBtc}}{character}

\item{\code{marginTradeCoeffVo}}{\link{InlineResponse20057MarginTradeCoeffVo}}

\item{\code{marginUserAssetVoList}}{list( \link{InlineResponse20019UserAssets} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20057$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20057$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20057$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20057$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20057$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20057$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057$new(
  email,
  marginLevel,
  totalAssetOfBtc,
  totalLiabilityOfBtc,
  totalNetAssetOfBtc,
  marginTradeCoeffVo,
  marginUserAssetVoList,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057$fromJSON(InlineResponse20057Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057$fromJSONString(InlineResponse20057Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
