% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20057_margin_trade_coeff_vo.R
\docType{class}
\name{InlineResponse20057MarginTradeCoeffVo}
\alias{InlineResponse20057MarginTradeCoeffVo}
\title{InlineResponse20057MarginTradeCoeffVo}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20057MarginTradeCoeffVo Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{forceLiquidationBar}}{character}

\item{\code{marginCallBar}}{character}

\item{\code{normalBar}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20057MarginTradeCoeffVo$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20057MarginTradeCoeffVo$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20057MarginTradeCoeffVo$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20057MarginTradeCoeffVo$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20057MarginTradeCoeffVo$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20057MarginTradeCoeffVo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057MarginTradeCoeffVo$new(
  forceLiquidationBar,
  marginCallBar,
  normalBar,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057MarginTradeCoeffVo$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057MarginTradeCoeffVo$fromJSON(
  InlineResponse20057MarginTradeCoeffVoJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057MarginTradeCoeffVo$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057MarginTradeCoeffVo$fromJSONString(
  InlineResponse20057MarginTradeCoeffVoJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20057MarginTradeCoeffVo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
