% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20061_sub_account_list.R
\docType{class}
\name{InlineResponse20061SubAccountList}
\alias{InlineResponse20061SubAccountList}
\title{InlineResponse20061SubAccountList}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20061SubAccountList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{character}

\item{\code{totalAssetOfBtc}}{character}

\item{\code{totalLiabilityOfBtc}}{character}

\item{\code{totalNetAssetOfBtc}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20061SubAccountList$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20061SubAccountList$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20061SubAccountList$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20061SubAccountList$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20061SubAccountList$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20061SubAccountList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20061SubAccountList$new(
  email,
  totalAssetOfBtc,
  totalLiabilityOfBtc,
  totalNetAssetOfBtc,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20061SubAccountList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20061SubAccountList$fromJSON(
  InlineResponse20061SubAccountListJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20061SubAccountList$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20061SubAccountList$fromJSONString(
  InlineResponse20061SubAccountListJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20061SubAccountList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
