% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20062.R
\docType{class}
\name{InlineResponse20062}
\alias{InlineResponse20062}
\title{InlineResponse20062}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20062 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entryPrice}}{character}

\item{\code{leverage}}{character}

\item{\code{maxNotional}}{character}

\item{\code{liquidationPrice}}{character}

\item{\code{markPrice}}{character}

\item{\code{positionAmount}}{character}

\item{\code{symbol}}{character}

\item{\code{unrealizedProfit}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20062$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20062$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20062$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20062$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20062$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20062$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20062$new(
  entryPrice,
  leverage,
  maxNotional,
  liquidationPrice,
  markPrice,
  positionAmount,
  symbol,
  unrealizedProfit,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20062$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20062$fromJSON(InlineResponse20062Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20062$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20062$fromJSONString(InlineResponse20062Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20062$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
