% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20063_assets.R
\docType{class}
\name{InlineResponse20063Assets}
\alias{InlineResponse20063Assets}
\title{InlineResponse20063Assets}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20063Assets Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{asset}}{character}

\item{\code{initialMargin}}{character}

\item{\code{maintenanceMargin}}{character}

\item{\code{marginBalance}}{character}

\item{\code{maxWithdrawAmount}}{character}

\item{\code{openOrderInitialMargin}}{character}

\item{\code{positionInitialMargin}}{character}

\item{\code{unrealizedProfit}}{character}

\item{\code{walletBalance}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20063Assets$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20063Assets$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20063Assets$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20063Assets$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20063Assets$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20063Assets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20063Assets$new(
  asset,
  initialMargin,
  maintenanceMargin,
  marginBalance,
  maxWithdrawAmount,
  openOrderInitialMargin,
  positionInitialMargin,
  unrealizedProfit,
  walletBalance,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20063Assets$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20063Assets$fromJSON(InlineResponse20063AssetsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20063Assets$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20063Assets$fromJSONString(InlineResponse20063AssetsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20063Assets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
