% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20064_sub_account_list.R
\docType{class}
\name{InlineResponse20064SubAccountList}
\alias{InlineResponse20064SubAccountList}
\title{InlineResponse20064SubAccountList}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20064SubAccountList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{character}

\item{\code{totalInitialMargin}}{character}

\item{\code{totalMaintenanceMargin}}{character}

\item{\code{totalMarginBalance}}{character}

\item{\code{totalOpenOrderInitialMargin}}{character}

\item{\code{totalPositionInitialMargin}}{character}

\item{\code{totalUnrealizedProfit}}{character}

\item{\code{totalWalletBalance}}{character}

\item{\code{asset}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20064SubAccountList$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20064SubAccountList$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20064SubAccountList$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20064SubAccountList$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20064SubAccountList$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20064SubAccountList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20064SubAccountList$new(
  email,
  totalInitialMargin,
  totalMaintenanceMargin,
  totalMarginBalance,
  totalOpenOrderInitialMargin,
  totalPositionInitialMargin,
  totalUnrealizedProfit,
  totalWalletBalance,
  asset,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20064SubAccountList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20064SubAccountList$fromJSON(
  InlineResponse20064SubAccountListJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20064SubAccountList$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20064SubAccountList$fromJSONString(
  InlineResponse20064SubAccountListJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20064SubAccountList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
