% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20067.R
\docType{class}
\name{InlineResponse20067}
\alias{InlineResponse20067}
\title{InlineResponse20067}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20067 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{counterParty}}{character}

\item{\code{email}}{character}

\item{\code{type}}{integer}

\item{\code{asset}}{character}

\item{\code{qty}}{character}

\item{\code{fromAccountType}}{character}

\item{\code{toAccountType}}{character}

\item{\code{status}}{character}

\item{\code{tranId}}{integer}

\item{\code{time}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20067$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20067$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20067$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20067$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20067$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20067$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20067$new(
  counterParty,
  email,
  type,
  asset,
  qty,
  fromAccountType,
  toAccountType,
  status,
  tranId,
  time,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20067$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20067$fromJSON(InlineResponse20067Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20067$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20067$fromJSONString(InlineResponse20067Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20067$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
