% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20068.R
\docType{class}
\name{InlineResponse20068}
\alias{InlineResponse20068}
\title{InlineResponse20068}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20068 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tranId}}{integer}

\item{\code{fromEmail}}{character}

\item{\code{toEmail}}{character}

\item{\code{asset}}{character}

\item{\code{amount}}{character}

\item{\code{fromAccountType}}{character}

\item{\code{toAccountType}}{character}

\item{\code{status}}{character}

\item{\code{createTimeStamp}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20068$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20068$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20068$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20068$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20068$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20068$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20068$new(
  tranId,
  fromEmail,
  toEmail,
  asset,
  amount,
  fromAccountType,
  toAccountType,
  status,
  createTimeStamp,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20068$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20068$fromJSON(InlineResponse20068Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20068$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20068$fromJSONString(InlineResponse20068Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20068$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
