% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response2007.R
\docType{class}
\name{InlineResponse2007}
\alias{InlineResponse2007}
\title{InlineResponse2007}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse2007 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{orderListId}}{integer}

\item{\code{contingencyType}}{character}

\item{\code{listStatusType}}{character}

\item{\code{listOrderStatus}}{character}

\item{\code{listClientOrderId}}{character}

\item{\code{transactionTime}}{integer}

\item{\code{symbol}}{character}

\item{\code{orders}}{list( \link{ApiV3OpenOrderListOrders} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse2007$new()}}
\item \href{#method-toJSON}{\code{InlineResponse2007$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse2007$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse2007$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse2007$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse2007$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2007$new(
  orderListId,
  contingencyType,
  listStatusType,
  listOrderStatus,
  listClientOrderId,
  transactionTime,
  symbol,
  orders,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2007$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2007$fromJSON(InlineResponse2007Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2007$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2007$fromJSONString(InlineResponse2007Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2007$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
