% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20078.R
\docType{class}
\name{InlineResponse20078}
\alias{InlineResponse20078}
\title{InlineResponse20078}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20078 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{annualInterestRate}}{character}

\item{\code{asset}}{character}

\item{\code{avgAnnualInterestRate}}{character}

\item{\code{canRedeem}}{character}

\item{\code{dailyInterestRate}}{character}

\item{\code{freeAmount}}{character}

\item{\code{freezeAmount}}{character}

\item{\code{lockedAmount}}{character}

\item{\code{productId}}{character}

\item{\code{productName}}{character}

\item{\code{redeemingAmount}}{character}

\item{\code{todayPurchasedAmount}}{character}

\item{\code{totalAmount}}{character}

\item{\code{totalInterest}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20078$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20078$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20078$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20078$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20078$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20078$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20078$new(
  annualInterestRate,
  asset,
  avgAnnualInterestRate,
  canRedeem,
  dailyInterestRate,
  freeAmount,
  freezeAmount,
  lockedAmount,
  productId,
  productName,
  redeemingAmount,
  todayPurchasedAmount,
  totalAmount,
  totalInterest,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20078$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20078$fromJSON(InlineResponse20078Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20078$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20078$fromJSONString(InlineResponse20078Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20078$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
