% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20079.R
\docType{class}
\name{InlineResponse20079}
\alias{InlineResponse20079}
\title{InlineResponse20079}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20079 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{asset}}{character}

\item{\code{displayPriority}}{integer}

\item{\code{duration}}{integer}

\item{\code{interestPerLot}}{character}

\item{\code{interestRate}}{character}

\item{\code{lotSize}}{character}

\item{\code{lotsLowLimit}}{integer}

\item{\code{lotsPurchased}}{integer}

\item{\code{lotsUpLimit}}{integer}

\item{\code{maxLotsPerUser}}{integer}

\item{\code{needKyc}}{character}

\item{\code{projectId}}{character}

\item{\code{projectName}}{character}

\item{\code{status}}{character}

\item{\code{type}}{character}

\item{\code{withAreaLimitation}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20079$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20079$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20079$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20079$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20079$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20079$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079$new(
  asset,
  displayPriority,
  duration,
  interestPerLot,
  interestRate,
  lotSize,
  lotsLowLimit,
  lotsPurchased,
  lotsUpLimit,
  maxLotsPerUser,
  needKyc,
  projectId,
  projectName,
  status,
  type,
  withAreaLimitation,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079$fromJSON(InlineResponse20079Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079$fromJSONString(InlineResponse20079Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
