% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20079_data.R
\docType{class}
\name{InlineResponse20079Data}
\alias{InlineResponse20079Data}
\title{InlineResponse20079Data}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20079Data Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{orderNo}}{character}

\item{\code{fiatCurrency}}{character}

\item{\code{indicatedAmount}}{character}

\item{\code{amount}}{character}

\item{\code{totalFee}}{character}

\item{\code{method}}{character}

\item{\code{status}}{character}

\item{\code{createTime}}{integer}

\item{\code{updateTime}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20079Data$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20079Data$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20079Data$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20079Data$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20079Data$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20079Data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079Data$new(
  orderNo,
  fiatCurrency,
  indicatedAmount,
  amount,
  totalFee,
  method,
  status,
  createTime,
  updateTime,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079Data$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079Data$fromJSON(InlineResponse20079DataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079Data$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079Data$fromJSONString(InlineResponse20079DataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20079Data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
