% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20081.R
\docType{class}
\name{InlineResponse20081}
\alias{InlineResponse20081}
\title{InlineResponse20081}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20081 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{asset}}{character}

\item{\code{canTransfer}}{character}

\item{\code{createTimestamp}}{integer}

\item{\code{duration}}{integer}

\item{\code{endTime}}{integer}

\item{\code{interest}}{character}

\item{\code{interestRate}}{character}

\item{\code{lot}}{integer}

\item{\code{positionId}}{integer}

\item{\code{principal}}{character}

\item{\code{projectId}}{character}

\item{\code{projectName}}{character}

\item{\code{purchaseTime}}{integer}

\item{\code{redeemDate}}{character}

\item{\code{startTime}}{integer}

\item{\code{status}}{character}

\item{\code{type}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20081$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20081$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20081$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20081$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20081$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20081$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20081$new(
  asset,
  canTransfer,
  createTimestamp,
  duration,
  endTime,
  interest,
  interestRate,
  lot,
  positionId,
  principal,
  projectId,
  projectName,
  purchaseTime,
  redeemDate,
  startTime,
  status,
  type,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20081$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20081$fromJSON(InlineResponse20081Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20081$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20081$fromJSONString(InlineResponse20081Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20081$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
