% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20082.R
\docType{class}
\name{InlineResponse20082}
\alias{InlineResponse20082}
\title{InlineResponse20082}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20082 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{positionAmountVos}}{list( \link{InlineResponse20082PositionAmountVos} )}

\item{\code{totalAmountInBTC}}{character}

\item{\code{totalAmountInUSDT}}{character}

\item{\code{totalFixedAmountInBTC}}{character}

\item{\code{totalFixedAmountInUSDT}}{character}

\item{\code{totalFlexibleInBTC}}{character}

\item{\code{totalFlexibleInUSDT}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20082$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20082$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20082$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20082$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20082$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20082$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082$new(
  positionAmountVos,
  totalAmountInBTC,
  totalAmountInUSDT,
  totalFixedAmountInBTC,
  totalFixedAmountInUSDT,
  totalFlexibleInBTC,
  totalFlexibleInUSDT,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082$fromJSON(InlineResponse20082Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082$fromJSONString(InlineResponse20082Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
