% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20082_position_amount_vos.R
\docType{class}
\name{InlineResponse20082PositionAmountVos}
\alias{InlineResponse20082PositionAmountVos}
\title{InlineResponse20082PositionAmountVos}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20082PositionAmountVos Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{amount}}{character}

\item{\code{amountInBTC}}{character}

\item{\code{amountInUSDT}}{character}

\item{\code{asset}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20082PositionAmountVos$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20082PositionAmountVos$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20082PositionAmountVos$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20082PositionAmountVos$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20082PositionAmountVos$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20082PositionAmountVos$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082PositionAmountVos$new(
  amount,
  amountInBTC,
  amountInUSDT,
  asset,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082PositionAmountVos$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082PositionAmountVos$fromJSON(
  InlineResponse20082PositionAmountVosJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082PositionAmountVos$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082PositionAmountVos$fromJSONString(
  InlineResponse20082PositionAmountVosJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20082PositionAmountVos$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
