% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20088_data_worker_datas.R
\docType{class}
\name{InlineResponse20088DataWorkerDatas}
\alias{InlineResponse20088DataWorkerDatas}
\title{InlineResponse20088DataWorkerDatas}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20088DataWorkerDatas Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{workerId}}{character}

\item{\code{workerName}}{character}

\item{\code{status}}{integer}

\item{\code{hashRate}}{integer}

\item{\code{dayHashRate}}{integer}

\item{\code{rejectRate}}{integer}

\item{\code{lastShareTime}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20088DataWorkerDatas$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20088DataWorkerDatas$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20088DataWorkerDatas$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20088DataWorkerDatas$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20088DataWorkerDatas$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20088DataWorkerDatas$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20088DataWorkerDatas$new(
  workerId,
  workerName,
  status,
  hashRate,
  dayHashRate,
  rejectRate,
  lastShareTime,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20088DataWorkerDatas$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20088DataWorkerDatas$fromJSON(
  InlineResponse20088DataWorkerDatasJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20088DataWorkerDatas$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20088DataWorkerDatas$fromJSONString(
  InlineResponse20088DataWorkerDatasJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20088DataWorkerDatas$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
