% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20091_data_config_details.R
\docType{class}
\name{InlineResponse20091DataConfigDetails}
\alias{InlineResponse20091DataConfigDetails}
\title{InlineResponse20091DataConfigDetails}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20091DataConfigDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{configId}}{integer}

\item{\code{poolUsername}}{character}

\item{\code{toPoolUsername}}{character}

\item{\code{algoName}}{character}

\item{\code{hashRate}}{integer}

\item{\code{startDay}}{integer}

\item{\code{endDay}}{integer}

\item{\code{status}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20091DataConfigDetails$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20091DataConfigDetails$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20091DataConfigDetails$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20091DataConfigDetails$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20091DataConfigDetails$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20091DataConfigDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20091DataConfigDetails$new(
  configId,
  poolUsername,
  toPoolUsername,
  algoName,
  hashRate,
  startDay,
  endDay,
  status,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20091DataConfigDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20091DataConfigDetails$fromJSON(
  InlineResponse20091DataConfigDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20091DataConfigDetails$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20091DataConfigDetails$fromJSONString(
  InlineResponse20091DataConfigDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20091DataConfigDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
