% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inline_response20092_data_profit_transfer_details.R
\docType{class}
\name{InlineResponse20092DataProfitTransferDetails}
\alias{InlineResponse20092DataProfitTransferDetails}
\title{InlineResponse20092DataProfitTransferDetails}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20092DataProfitTransferDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{poolUsername}}{character}

\item{\code{toPoolUsername}}{character}

\item{\code{algoName}}{character}

\item{\code{hashRate}}{integer}

\item{\code{day}}{integer}

\item{\code{amount}}{numeric}

\item{\code{coinName}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20092DataProfitTransferDetails$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20092DataProfitTransferDetails$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20092DataProfitTransferDetails$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20092DataProfitTransferDetails$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20092DataProfitTransferDetails$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20092DataProfitTransferDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20092DataProfitTransferDetails$new(
  poolUsername,
  toPoolUsername,
  algoName,
  hashRate,
  day,
  amount,
  coinName,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20092DataProfitTransferDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20092DataProfitTransferDetails$fromJSON(
  InlineResponse20092DataProfitTransferDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20092DataProfitTransferDetails$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20092DataProfitTransferDetails$fromJSONString(
  InlineResponse20092DataProfitTransferDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20092DataProfitTransferDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
