% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20094_hashrate_datas.R
\docType{class}
\name{InlineResponse20094HashrateDatas}
\alias{InlineResponse20094HashrateDatas}
\title{InlineResponse20094HashrateDatas}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20094HashrateDatas Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time}}{integer}

\item{\code{hashrate}}{character}

\item{\code{reject}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20094HashrateDatas$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20094HashrateDatas$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20094HashrateDatas$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20094HashrateDatas$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20094HashrateDatas$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20094HashrateDatas$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20094HashrateDatas$new(time, hashrate, reject, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20094HashrateDatas$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20094HashrateDatas$fromJSON(InlineResponse20094HashrateDatasJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20094HashrateDatas$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20094HashrateDatas$fromJSONString(
  InlineResponse20094HashrateDatasJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20094HashrateDatas$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
