% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20095_data_profit_today.R
\docType{class}
\name{InlineResponse20095DataProfitToday}
\alias{InlineResponse20095DataProfitToday}
\title{InlineResponse20095DataProfitToday}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20095DataProfitToday Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{BTC}}{character}

\item{\code{BSV}}{character}

\item{\code{BCH}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20095DataProfitToday$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20095DataProfitToday$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20095DataProfitToday$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20095DataProfitToday$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20095DataProfitToday$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20095DataProfitToday$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20095DataProfitToday$new(BTC, BSV, BCH, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20095DataProfitToday$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20095DataProfitToday$fromJSON(
  InlineResponse20095DataProfitTodayJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20095DataProfitToday$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20095DataProfitToday$fromJSONString(
  InlineResponse20095DataProfitTodayJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20095DataProfitToday$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
