% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20096_data_account_profits.R
\docType{class}
\name{InlineResponse20096DataAccountProfits}
\alias{InlineResponse20096DataAccountProfits}
\title{InlineResponse20096DataAccountProfits}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20096DataAccountProfits Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time}}{integer}

\item{\code{type}}{integer}

\item{\code{hashTransfer}}{integer}

\item{\code{transferAmount}}{numeric}

\item{\code{dayHashRate}}{integer}

\item{\code{profitAmount}}{numeric}

\item{\code{coinName}}{character}

\item{\code{status}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20096DataAccountProfits$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20096DataAccountProfits$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20096DataAccountProfits$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20096DataAccountProfits$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20096DataAccountProfits$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20096DataAccountProfits$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20096DataAccountProfits$new(
  time,
  type,
  hashTransfer,
  transferAmount,
  dayHashRate,
  profitAmount,
  coinName,
  status,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20096DataAccountProfits$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20096DataAccountProfits$fromJSON(
  InlineResponse20096DataAccountProfitsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20096DataAccountProfits$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20096DataAccountProfits$fromJSONString(
  InlineResponse20096DataAccountProfitsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20096DataAccountProfits$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
