% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20097_data_other_profits.R
\docType{class}
\name{InlineResponse20097DataOtherProfits}
\alias{InlineResponse20097DataOtherProfits}
\title{InlineResponse20097DataOtherProfits}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20097DataOtherProfits Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time}}{integer}

\item{\code{coinName}}{character}

\item{\code{type}}{integer}

\item{\code{profitAmount}}{numeric}

\item{\code{status}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20097DataOtherProfits$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20097DataOtherProfits$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20097DataOtherProfits$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20097DataOtherProfits$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20097DataOtherProfits$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20097DataOtherProfits$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097DataOtherProfits$new(
  time,
  coinName,
  type,
  profitAmount,
  status,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097DataOtherProfits$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097DataOtherProfits$fromJSON(
  InlineResponse20097DataOtherProfitsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097DataOtherProfits$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097DataOtherProfits$fromJSONString(
  InlineResponse20097DataOtherProfitsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097DataOtherProfits$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
