% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolated_margin_account_info_base_asset.R
\docType{class}
\name{IsolatedMarginAccountInfoBaseAsset}
\alias{IsolatedMarginAccountInfoBaseAsset}
\title{IsolatedMarginAccountInfoBaseAsset}
\format{
An \code{R6Class} generator object
}
\description{
IsolatedMarginAccountInfoBaseAsset Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{asset}}{character}

\item{\code{borrowEnabled}}{character}

\item{\code{borrowed}}{character}

\item{\code{free}}{character}

\item{\code{interest}}{character}

\item{\code{locked}}{character}

\item{\code{netAsset}}{character}

\item{\code{netAssetOfBtc}}{character}

\item{\code{repayEnabled}}{character}

\item{\code{totalAsset}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IsolatedMarginAccountInfoBaseAsset$new()}}
\item \href{#method-toJSON}{\code{IsolatedMarginAccountInfoBaseAsset$toJSON()}}
\item \href{#method-fromJSON}{\code{IsolatedMarginAccountInfoBaseAsset$fromJSON()}}
\item \href{#method-toJSONString}{\code{IsolatedMarginAccountInfoBaseAsset$toJSONString()}}
\item \href{#method-fromJSONString}{\code{IsolatedMarginAccountInfoBaseAsset$fromJSONString()}}
\item \href{#method-clone}{\code{IsolatedMarginAccountInfoBaseAsset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsolatedMarginAccountInfoBaseAsset$new(
  asset,
  borrowEnabled,
  borrowed,
  free,
  interest,
  locked,
  netAsset,
  netAssetOfBtc,
  repayEnabled,
  totalAsset,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsolatedMarginAccountInfoBaseAsset$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsolatedMarginAccountInfoBaseAsset$fromJSON(
  IsolatedMarginAccountInfoBaseAssetJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsolatedMarginAccountInfoBaseAsset$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsolatedMarginAccountInfoBaseAsset$fromJSONString(
  IsolatedMarginAccountInfoBaseAssetJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsolatedMarginAccountInfoBaseAsset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
