% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_oco_order_order_reports.R
\docType{class}
\name{MarginOcoOrderOrderReports}
\alias{MarginOcoOrderOrderReports}
\title{MarginOcoOrderOrderReports}
\format{
An \code{R6Class} generator object
}
\description{
MarginOcoOrderOrderReports Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{symbol}}{character}

\item{\code{origClientOrderId}}{character}

\item{\code{orderId}}{integer}

\item{\code{orderListId}}{integer}

\item{\code{clientOrderId}}{character}

\item{\code{price}}{character}

\item{\code{origQty}}{character}

\item{\code{executedQty}}{character}

\item{\code{cummulativeQuoteQty}}{character}

\item{\code{status}}{character}

\item{\code{timeInForce}}{character}

\item{\code{type}}{character}

\item{\code{side}}{character}

\item{\code{stopPrice}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MarginOcoOrderOrderReports$new()}}
\item \href{#method-toJSON}{\code{MarginOcoOrderOrderReports$toJSON()}}
\item \href{#method-fromJSON}{\code{MarginOcoOrderOrderReports$fromJSON()}}
\item \href{#method-toJSONString}{\code{MarginOcoOrderOrderReports$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MarginOcoOrderOrderReports$fromJSONString()}}
\item \href{#method-clone}{\code{MarginOcoOrderOrderReports$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOcoOrderOrderReports$new(
  symbol,
  origClientOrderId,
  orderId,
  orderListId,
  clientOrderId,
  price,
  origQty,
  executedQty,
  cummulativeQuoteQty,
  status,
  timeInForce,
  type,
  side,
  stopPrice,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOcoOrderOrderReports$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOcoOrderOrderReports$fromJSON(MarginOcoOrderOrderReportsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOcoOrderOrderReports$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOcoOrderOrderReports$fromJSONString(MarginOcoOrderOrderReportsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOcoOrderOrderReports$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
