% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_order_detail.R
\docType{class}
\name{MarginOrderDetail}
\alias{MarginOrderDetail}
\title{MarginOrderDetail}
\format{
An \code{R6Class} generator object
}
\description{
MarginOrderDetail Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{clientOrderId}}{character}

\item{\code{cummulativeQuoteQty}}{character}

\item{\code{executedQty}}{character}

\item{\code{icebergQty}}{character}

\item{\code{isWorking}}{character}

\item{\code{orderId}}{integer}

\item{\code{origQty}}{character}

\item{\code{price}}{character}

\item{\code{side}}{character}

\item{\code{status}}{character}

\item{\code{stopPrice}}{character}

\item{\code{symbol}}{character}

\item{\code{isIsolated}}{character}

\item{\code{time}}{integer}

\item{\code{timeInForce}}{character}

\item{\code{type}}{character}

\item{\code{updateTime}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MarginOrderDetail$new()}}
\item \href{#method-toJSON}{\code{MarginOrderDetail$toJSON()}}
\item \href{#method-fromJSON}{\code{MarginOrderDetail$fromJSON()}}
\item \href{#method-toJSONString}{\code{MarginOrderDetail$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MarginOrderDetail$fromJSONString()}}
\item \href{#method-clone}{\code{MarginOrderDetail$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderDetail$new(
  clientOrderId,
  cummulativeQuoteQty,
  executedQty,
  icebergQty,
  isWorking,
  orderId,
  origQty,
  price,
  side,
  status,
  stopPrice,
  symbol,
  isIsolated,
  time,
  timeInForce,
  type,
  updateTime,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderDetail$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderDetail$fromJSON(MarginOrderDetailJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderDetail$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderDetail$fromJSONString(MarginOrderDetailJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderDetail$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
