% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_order_response_full.R
\docType{class}
\name{MarginOrderResponseFull}
\alias{MarginOrderResponseFull}
\title{MarginOrderResponseFull}
\format{
An \code{R6Class} generator object
}
\description{
MarginOrderResponseFull Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{symbol}}{character}

\item{\code{orderId}}{integer}

\item{\code{clientOrderId}}{character}

\item{\code{transactTime}}{integer}

\item{\code{price}}{character}

\item{\code{origQty}}{character}

\item{\code{executedQty}}{character}

\item{\code{cummulativeQuoteQty}}{character}

\item{\code{status}}{character}

\item{\code{timeInForce}}{character}

\item{\code{type}}{character}

\item{\code{side}}{character}

\item{\code{marginBuyBorrowAmount}}{numeric}

\item{\code{marginBuyBorrowAsset}}{character}

\item{\code{isIsolated}}{character}

\item{\code{fills}}{list( \link{OrderResponseFullFills} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MarginOrderResponseFull$new()}}
\item \href{#method-toJSON}{\code{MarginOrderResponseFull$toJSON()}}
\item \href{#method-fromJSON}{\code{MarginOrderResponseFull$fromJSON()}}
\item \href{#method-toJSONString}{\code{MarginOrderResponseFull$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MarginOrderResponseFull$fromJSONString()}}
\item \href{#method-clone}{\code{MarginOrderResponseFull$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderResponseFull$new(
  symbol,
  orderId,
  clientOrderId,
  transactTime,
  price,
  origQty,
  executedQty,
  cummulativeQuoteQty,
  status,
  timeInForce,
  type,
  side,
  marginBuyBorrowAmount,
  marginBuyBorrowAsset,
  isIsolated,
  fills,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderResponseFull$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderResponseFull$fromJSON(MarginOrderResponseFullJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderResponseFull$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderResponseFull$fromJSONString(MarginOrderResponseFullJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginOrderResponseFull$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
