% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_trade.R
\docType{class}
\name{MarginTrade}
\alias{MarginTrade}
\title{MarginTrade}
\format{
An \code{R6Class} generator object
}
\description{
MarginTrade Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{commission}}{character}

\item{\code{commissionAsset}}{character}

\item{\code{id}}{integer}

\item{\code{isBestMatch}}{character}

\item{\code{isBuyer}}{character}

\item{\code{isMaker}}{character}

\item{\code{orderId}}{integer}

\item{\code{price}}{character}

\item{\code{qty}}{character}

\item{\code{symbol}}{character}

\item{\code{isIsolated}}{character}

\item{\code{time}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MarginTrade$new()}}
\item \href{#method-toJSON}{\code{MarginTrade$toJSON()}}
\item \href{#method-fromJSON}{\code{MarginTrade$fromJSON()}}
\item \href{#method-toJSONString}{\code{MarginTrade$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MarginTrade$fromJSONString()}}
\item \href{#method-clone}{\code{MarginTrade$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTrade$new(
  commission,
  commissionAsset,
  id,
  isBestMatch,
  isBuyer,
  isMaker,
  orderId,
  price,
  qty,
  symbol,
  isIsolated,
  time,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTrade$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTrade$fromJSON(MarginTradeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTrade$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTrade$fromJSONString(MarginTradeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTrade$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
