% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_transfer_details_rows.R
\docType{class}
\name{MarginTransferDetailsRows}
\alias{MarginTransferDetailsRows}
\title{MarginTransferDetailsRows}
\format{
An \code{R6Class} generator object
}
\description{
MarginTransferDetailsRows Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{amount}}{character}

\item{\code{asset}}{character}

\item{\code{status}}{character}

\item{\code{timestamp}}{integer}

\item{\code{txId}}{integer}

\item{\code{transFrom}}{character}

\item{\code{transTo}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MarginTransferDetailsRows$new()}}
\item \href{#method-toJSON}{\code{MarginTransferDetailsRows$toJSON()}}
\item \href{#method-fromJSON}{\code{MarginTransferDetailsRows$fromJSON()}}
\item \href{#method-toJSONString}{\code{MarginTransferDetailsRows$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MarginTransferDetailsRows$fromJSONString()}}
\item \href{#method-clone}{\code{MarginTransferDetailsRows$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTransferDetailsRows$new(
  amount,
  asset,
  status,
  timestamp,
  txId,
  transFrom,
  transTo,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTransferDetailsRows$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTransferDetailsRows$fromJSON(MarginTransferDetailsRowsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTransferDetailsRows$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTransferDetailsRows$fromJSONString(MarginTransferDetailsRowsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginTransferDetailsRows$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
