% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nft_api.R
\docType{class}
\name{NFTApi}
\alias{NFTApi}
\title{NFT operations}
\format{
An \code{R6Class} generator object
}
\description{
binanceRapi.NFT
}
\section{Methods}{

\describe{
\strong{ SapiV1NftHistoryDepositGet } \emph{ Get NFT Deposit History(USER_DATA) }
- The max interval between startTime and endTime is 90 days. - If startTime and endTime are not sent, the recent 7 days&#39; data will be returned.  Weight(UID): 3000

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } limit integer
\item \emph{ @param } page integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse200124} \cr


\item status code : 200 | NFT Deposit History

\item return type : InlineResponse200124 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1NftHistoryTransactionsGet } \emph{ Get NFT Transaction History (USER_DATA) }
- The max interval between startTime and endTime is 90 days. - If startTime and endTime are not sent, the recent 7 days&#39; data will be returned.  Weight(UID): 3000

\itemize{
\item \emph{ @param } order.type integer
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } limit integer
\item \emph{ @param } page integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse200123} \cr


\item status code : 200 | NFT Transaction History

\item return type : InlineResponse200123 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1NftHistoryWithdrawGet } \emph{ Get NFT Withdraw History (USER_DATA) }
- The max interval between startTime and endTime is 90 days. - If startTime and endTime are not sent, the recent 7 days&#39; data will be returned.  Weight(UID): 3000

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } limit integer
\item \emph{ @param } page integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse200125} \cr


\item status code : 200 | NFT Withdraw History

\item return type : InlineResponse200125 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1NftUserGetAssetGet } \emph{ Get NFT Asset (USER_DATA) }
Weight(UID): 3000

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } limit integer
\item \emph{ @param } page integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse200126} \cr


\item status code : 200 | Asset Information

\item return type : InlineResponse200126 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  SapiV1NftHistoryDepositGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.limit <- 50 # integer | Default 50, Max 50
var.page <- 1 # integer | Default 1
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get NFT Deposit History(USER_DATA)
api.instance <- NFTApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1NftHistoryDepositGet(var.timestamp, var.signature, start.time=var.start.time, end.time=var.end.time, limit=var.limit, page=var.page, recv.window=var.recv.window)


####################  SapiV1NftHistoryTransactionsGet  ####################

library(binanceRapi)
var.order.type <- 1 # integer | 0: purchase order, 1: sell order, 2: royalty income, 3: primary market order, 4: mint fee
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.limit <- 50 # integer | Default 50, Max 50
var.page <- 1 # integer | Default 1
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get NFT Transaction History (USER_DATA)
api.instance <- NFTApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1NftHistoryTransactionsGet(var.order.type, var.timestamp, var.signature, start.time=var.start.time, end.time=var.end.time, limit=var.limit, page=var.page, recv.window=var.recv.window)


####################  SapiV1NftHistoryWithdrawGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.limit <- 50 # integer | Default 50, Max 50
var.page <- 1 # integer | Default 1
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get NFT Withdraw History (USER_DATA)
api.instance <- NFTApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1NftHistoryWithdrawGet(var.timestamp, var.signature, start.time=var.start.time, end.time=var.end.time, limit=var.limit, page=var.page, recv.window=var.recv.window)


####################  SapiV1NftUserGetAssetGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.limit <- 50 # integer | Default 50, Max 50
var.page <- 1 # integer | Default 1
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get NFT Asset (USER_DATA)
api.instance <- NFTApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1NftUserGetAssetGet(var.timestamp, var.signature, limit=var.limit, page=var.page, recv.window=var.recv.window)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NFTApi$new()}}
\item \href{#method-SapiV1NftHistoryDepositGet}{\code{NFTApi$SapiV1NftHistoryDepositGet()}}
\item \href{#method-SapiV1NftHistoryDepositGetWithHttpInfo}{\code{NFTApi$SapiV1NftHistoryDepositGetWithHttpInfo()}}
\item \href{#method-SapiV1NftHistoryTransactionsGet}{\code{NFTApi$SapiV1NftHistoryTransactionsGet()}}
\item \href{#method-SapiV1NftHistoryTransactionsGetWithHttpInfo}{\code{NFTApi$SapiV1NftHistoryTransactionsGetWithHttpInfo()}}
\item \href{#method-SapiV1NftHistoryWithdrawGet}{\code{NFTApi$SapiV1NftHistoryWithdrawGet()}}
\item \href{#method-SapiV1NftHistoryWithdrawGetWithHttpInfo}{\code{NFTApi$SapiV1NftHistoryWithdrawGetWithHttpInfo()}}
\item \href{#method-SapiV1NftUserGetAssetGet}{\code{NFTApi$SapiV1NftUserGetAssetGet()}}
\item \href{#method-SapiV1NftUserGetAssetGetWithHttpInfo}{\code{NFTApi$SapiV1NftUserGetAssetGetWithHttpInfo()}}
\item \href{#method-clone}{\code{NFTApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFTApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1NftHistoryDepositGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1NftHistoryDepositGet}{}}}
\subsection{Method \code{SapiV1NftHistoryDepositGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFTApi$SapiV1NftHistoryDepositGet(
  timestamp,
  signature,
  start.time = NULL,
  end.time = NULL,
  limit = NULL,
  page = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1NftHistoryDepositGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1NftHistoryDepositGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1NftHistoryDepositGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFTApi$SapiV1NftHistoryDepositGetWithHttpInfo(
  timestamp,
  signature,
  start.time = NULL,
  end.time = NULL,
  limit = NULL,
  page = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1NftHistoryTransactionsGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1NftHistoryTransactionsGet}{}}}
\subsection{Method \code{SapiV1NftHistoryTransactionsGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFTApi$SapiV1NftHistoryTransactionsGet(
  order.type,
  timestamp,
  signature,
  start.time = NULL,
  end.time = NULL,
  limit = NULL,
  page = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1NftHistoryTransactionsGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1NftHistoryTransactionsGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1NftHistoryTransactionsGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFTApi$SapiV1NftHistoryTransactionsGetWithHttpInfo(
  order.type,
  timestamp,
  signature,
  start.time = NULL,
  end.time = NULL,
  limit = NULL,
  page = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1NftHistoryWithdrawGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1NftHistoryWithdrawGet}{}}}
\subsection{Method \code{SapiV1NftHistoryWithdrawGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFTApi$SapiV1NftHistoryWithdrawGet(
  timestamp,
  signature,
  start.time = NULL,
  end.time = NULL,
  limit = NULL,
  page = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1NftHistoryWithdrawGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1NftHistoryWithdrawGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1NftHistoryWithdrawGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFTApi$SapiV1NftHistoryWithdrawGetWithHttpInfo(
  timestamp,
  signature,
  start.time = NULL,
  end.time = NULL,
  limit = NULL,
  page = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1NftUserGetAssetGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1NftUserGetAssetGet}{}}}
\subsection{Method \code{SapiV1NftUserGetAssetGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFTApi$SapiV1NftUserGetAssetGet(
  timestamp,
  signature,
  limit = NULL,
  page = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1NftUserGetAssetGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1NftUserGetAssetGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1NftUserGetAssetGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFTApi$SapiV1NftUserGetAssetGetWithHttpInfo(
  timestamp,
  signature,
  limit = NULL,
  page = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFTApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
