% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oco_order.R
\docType{class}
\name{OcoOrder}
\alias{OcoOrder}
\title{OcoOrder}
\format{
An \code{R6Class} generator object
}
\description{
OcoOrder Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{orderListId}}{integer}

\item{\code{contingencyType}}{character}

\item{\code{listStatusType}}{character}

\item{\code{listOrderStatus}}{character}

\item{\code{listClientOrderId}}{character}

\item{\code{transactionTime}}{integer}

\item{\code{symbol}}{character}

\item{\code{orders}}{list( \link{InlineResponse20020Orders} )}

\item{\code{orderReports}}{list( \link{OcoOrderOrderReports} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OcoOrder$new()}}
\item \href{#method-toJSON}{\code{OcoOrder$toJSON()}}
\item \href{#method-fromJSON}{\code{OcoOrder$fromJSON()}}
\item \href{#method-toJSONString}{\code{OcoOrder$toJSONString()}}
\item \href{#method-fromJSONString}{\code{OcoOrder$fromJSONString()}}
\item \href{#method-clone}{\code{OcoOrder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OcoOrder$new(
  orderListId,
  contingencyType,
  listStatusType,
  listOrderStatus,
  listClientOrderId,
  transactionTime,
  symbol,
  orders,
  orderReports,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OcoOrder$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OcoOrder$fromJSON(OcoOrderJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OcoOrder$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OcoOrder$fromJSONString(OcoOrderJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OcoOrder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
