% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_details.R
\docType{class}
\name{OrderDetails}
\alias{OrderDetails}
\title{OrderDetails}
\format{
An \code{R6Class} generator object
}
\description{
OrderDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{symbol}}{character}

\item{\code{orderId}}{integer}

\item{\code{orderListId}}{integer}

\item{\code{clientOrderId}}{character}

\item{\code{price}}{character}

\item{\code{origQty}}{character}

\item{\code{executedQty}}{character}

\item{\code{cummulativeQuoteQty}}{character}

\item{\code{status}}{character}

\item{\code{timeInForce}}{character}

\item{\code{type}}{character}

\item{\code{side}}{character}

\item{\code{stopPrice}}{character}

\item{\code{icebergQty}}{character}

\item{\code{time}}{integer}

\item{\code{updateTime}}{integer}

\item{\code{isWorking}}{character}

\item{\code{origQuoteOrderQty}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OrderDetails$new()}}
\item \href{#method-toJSON}{\code{OrderDetails$toJSON()}}
\item \href{#method-fromJSON}{\code{OrderDetails$fromJSON()}}
\item \href{#method-toJSONString}{\code{OrderDetails$toJSONString()}}
\item \href{#method-fromJSONString}{\code{OrderDetails$fromJSONString()}}
\item \href{#method-clone}{\code{OrderDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderDetails$new(
  symbol,
  orderId,
  orderListId,
  clientOrderId,
  price,
  origQty,
  executedQty,
  cummulativeQuoteQty,
  status,
  timeInForce,
  type,
  side,
  stopPrice,
  icebergQty,
  time,
  updateTime,
  isWorking,
  origQuoteOrderQty,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderDetails$fromJSON(OrderDetailsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderDetails$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderDetails$fromJSONString(OrderDetailsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
