% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_response_result.R
\docType{class}
\name{OrderResponseResult}
\alias{OrderResponseResult}
\title{OrderResponseResult}
\format{
An \code{R6Class} generator object
}
\description{
OrderResponseResult Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{symbol}}{character}

\item{\code{orderId}}{integer}

\item{\code{orderListId}}{integer}

\item{\code{clientOrderId}}{character}

\item{\code{transactTime}}{integer}

\item{\code{price}}{character}

\item{\code{origQty}}{character}

\item{\code{executedQty}}{character}

\item{\code{cummulativeQuoteQty}}{character}

\item{\code{status}}{character}

\item{\code{timeInForce}}{character}

\item{\code{type}}{character}

\item{\code{side}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OrderResponseResult$new()}}
\item \href{#method-toJSON}{\code{OrderResponseResult$toJSON()}}
\item \href{#method-fromJSON}{\code{OrderResponseResult$fromJSON()}}
\item \href{#method-toJSONString}{\code{OrderResponseResult$toJSONString()}}
\item \href{#method-fromJSONString}{\code{OrderResponseResult$fromJSONString()}}
\item \href{#method-clone}{\code{OrderResponseResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderResponseResult$new(
  symbol,
  orderId,
  orderListId,
  clientOrderId,
  transactTime,
  price,
  origQty,
  executedQty,
  cummulativeQuoteQty,
  status,
  timeInForce,
  type,
  side,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderResponseResult$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderResponseResult$fromJSON(OrderResponseResultJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderResponseResult$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderResponseResult$fromJSONString(OrderResponseResultJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderResponseResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
