% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_api.R
\docType{class}
\name{RebateApi}
\alias{RebateApi}
\title{Rebate operations}
\format{
An \code{R6Class} generator object
}
\description{
binanceRapi.Rebate
}
\section{Methods}{

\describe{
\strong{ SapiV1RebateTaxQueryGet } \emph{ Get Spot Rebate History Records (USER_DATA) }
- The max interval between startTime and endTime is 90 days. - If startTime and endTime are not sent, the recent 7 days&#39; data will be returned. - The earliest startTime is supported on June 10, 2020  Weight(UID): 3000

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } page integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse200122} \cr


\item status code : 200 | Rebate History

\item return type : InlineResponse200122 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  SapiV1RebateTaxQueryGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.page <- 1 # integer | default 1
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get Spot Rebate History Records (USER_DATA)
api.instance <- RebateApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1RebateTaxQueryGet(var.timestamp, var.signature, start.time=var.start.time, end.time=var.end.time, page=var.page, recv.window=var.recv.window)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RebateApi$new()}}
\item \href{#method-SapiV1RebateTaxQueryGet}{\code{RebateApi$SapiV1RebateTaxQueryGet()}}
\item \href{#method-SapiV1RebateTaxQueryGetWithHttpInfo}{\code{RebateApi$SapiV1RebateTaxQueryGetWithHttpInfo()}}
\item \href{#method-clone}{\code{RebateApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RebateApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1RebateTaxQueryGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1RebateTaxQueryGet}{}}}
\subsection{Method \code{SapiV1RebateTaxQueryGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RebateApi$SapiV1RebateTaxQueryGet(
  timestamp,
  signature,
  start.time = NULL,
  end.time = NULL,
  page = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1RebateTaxQueryGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1RebateTaxQueryGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1RebateTaxQueryGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RebateApi$SapiV1RebateTaxQueryGetWithHttpInfo(
  timestamp,
  signature,
  start.time = NULL,
  end.time = NULL,
  page = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RebateApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
