% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapi_v1_blvt_token_info_current_baskets.R
\docType{class}
\name{SapiV1BlvtTokenInfoCurrentBaskets}
\alias{SapiV1BlvtTokenInfoCurrentBaskets}
\title{SapiV1BlvtTokenInfoCurrentBaskets}
\format{
An \code{R6Class} generator object
}
\description{
SapiV1BlvtTokenInfoCurrentBaskets Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{symbol}}{character}

\item{\code{amount}}{character}

\item{\code{notionalValue}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SapiV1BlvtTokenInfoCurrentBaskets$new()}}
\item \href{#method-toJSON}{\code{SapiV1BlvtTokenInfoCurrentBaskets$toJSON()}}
\item \href{#method-fromJSON}{\code{SapiV1BlvtTokenInfoCurrentBaskets$fromJSON()}}
\item \href{#method-toJSONString}{\code{SapiV1BlvtTokenInfoCurrentBaskets$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SapiV1BlvtTokenInfoCurrentBaskets$fromJSONString()}}
\item \href{#method-clone}{\code{SapiV1BlvtTokenInfoCurrentBaskets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BlvtTokenInfoCurrentBaskets$new(symbol, amount, notionalValue, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BlvtTokenInfoCurrentBaskets$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BlvtTokenInfoCurrentBaskets$fromJSON(
  SapiV1BlvtTokenInfoCurrentBasketsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BlvtTokenInfoCurrentBaskets$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BlvtTokenInfoCurrentBaskets$fromJSONString(
  SapiV1BlvtTokenInfoCurrentBasketsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BlvtTokenInfoCurrentBaskets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
