% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapi_v1_bswap_liquidity_share_asset.R
\docType{class}
\name{SapiV1BswapLiquidityShareAsset}
\alias{SapiV1BswapLiquidityShareAsset}
\title{SapiV1BswapLiquidityShareAsset}
\format{
An \code{R6Class} generator object
}
\description{
SapiV1BswapLiquidityShareAsset Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{BUSD}}{numeric}

\item{\code{USDT}}{numeric}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SapiV1BswapLiquidityShareAsset$new()}}
\item \href{#method-toJSON}{\code{SapiV1BswapLiquidityShareAsset$toJSON()}}
\item \href{#method-fromJSON}{\code{SapiV1BswapLiquidityShareAsset$fromJSON()}}
\item \href{#method-toJSONString}{\code{SapiV1BswapLiquidityShareAsset$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SapiV1BswapLiquidityShareAsset$fromJSONString()}}
\item \href{#method-clone}{\code{SapiV1BswapLiquidityShareAsset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapLiquidityShareAsset$new(BUSD, USDT, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapLiquidityShareAsset$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapLiquidityShareAsset$fromJSON(SapiV1BswapLiquidityShareAssetJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapLiquidityShareAsset$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapLiquidityShareAsset$fromJSONString(
  SapiV1BswapLiquidityShareAssetJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapLiquidityShareAsset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
