% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapi_v1_bswap_pool_configure_liquidity.R
\docType{class}
\name{SapiV1BswapPoolConfigureLiquidity}
\alias{SapiV1BswapPoolConfigureLiquidity}
\title{SapiV1BswapPoolConfigureLiquidity}
\format{
An \code{R6Class} generator object
}
\description{
SapiV1BswapPoolConfigureLiquidity Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{constantA}}{integer}

\item{\code{minRedeemShare}}{numeric}

\item{\code{slippageTolerance}}{numeric}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SapiV1BswapPoolConfigureLiquidity$new()}}
\item \href{#method-toJSON}{\code{SapiV1BswapPoolConfigureLiquidity$toJSON()}}
\item \href{#method-fromJSON}{\code{SapiV1BswapPoolConfigureLiquidity$fromJSON()}}
\item \href{#method-toJSONString}{\code{SapiV1BswapPoolConfigureLiquidity$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SapiV1BswapPoolConfigureLiquidity$fromJSONString()}}
\item \href{#method-clone}{\code{SapiV1BswapPoolConfigureLiquidity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapPoolConfigureLiquidity$new(
  constantA,
  minRedeemShare,
  slippageTolerance,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapPoolConfigureLiquidity$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapPoolConfigureLiquidity$fromJSON(
  SapiV1BswapPoolConfigureLiquidityJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapPoolConfigureLiquidity$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapPoolConfigureLiquidity$fromJSONString(
  SapiV1BswapPoolConfigureLiquidityJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1BswapPoolConfigureLiquidity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
