% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapi_v1_capital_config_getall_network_list.R
\docType{class}
\name{SapiV1CapitalConfigGetallNetworkList}
\alias{SapiV1CapitalConfigGetallNetworkList}
\title{SapiV1CapitalConfigGetallNetworkList}
\format{
An \code{R6Class} generator object
}
\description{
SapiV1CapitalConfigGetallNetworkList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{addressRegex}}{character}

\item{\code{coin}}{character}

\item{\code{depositDesc}}{character}

\item{\code{depositEnable}}{character}

\item{\code{isDefault}}{character}

\item{\code{memoRegex}}{character}

\item{\code{minConfirm}}{integer}

\item{\code{name}}{character}

\item{\code{resetAddressStatus}}{character}

\item{\code{specialTips}}{character}

\item{\code{unLockConfirm}}{integer}

\item{\code{withdrawDesc}}{character}

\item{\code{withdrawEnable}}{character}

\item{\code{withdrawFee}}{character}

\item{\code{withdrawIntegerMultiple}}{character}

\item{\code{withdrawMax}}{character}

\item{\code{withdrawMin}}{character}

\item{\code{sameAddress}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SapiV1CapitalConfigGetallNetworkList$new()}}
\item \href{#method-toJSON}{\code{SapiV1CapitalConfigGetallNetworkList$toJSON()}}
\item \href{#method-fromJSON}{\code{SapiV1CapitalConfigGetallNetworkList$fromJSON()}}
\item \href{#method-toJSONString}{\code{SapiV1CapitalConfigGetallNetworkList$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SapiV1CapitalConfigGetallNetworkList$fromJSONString()}}
\item \href{#method-clone}{\code{SapiV1CapitalConfigGetallNetworkList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1CapitalConfigGetallNetworkList$new(
  addressRegex,
  coin,
  depositDesc,
  depositEnable,
  isDefault,
  memoRegex,
  minConfirm,
  name,
  resetAddressStatus,
  specialTips,
  unLockConfirm,
  withdrawDesc,
  withdrawEnable,
  withdrawFee,
  withdrawIntegerMultiple,
  withdrawMax,
  withdrawMin,
  sameAddress,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1CapitalConfigGetallNetworkList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1CapitalConfigGetallNetworkList$fromJSON(
  SapiV1CapitalConfigGetallNetworkListJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1CapitalConfigGetallNetworkList$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1CapitalConfigGetallNetworkList$fromJSONString(
  SapiV1CapitalConfigGetallNetworkListJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1CapitalConfigGetallNetworkList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
