% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapi_v1_margin_isolated_margin_data_data.R
\docType{class}
\name{SapiV1MarginIsolatedMarginDataData}
\alias{SapiV1MarginIsolatedMarginDataData}
\title{SapiV1MarginIsolatedMarginDataData}
\format{
An \code{R6Class} generator object
}
\description{
SapiV1MarginIsolatedMarginDataData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coin}}{character [optional]}

\item{\code{dailyInterest}}{character [optional]}

\item{\code{borrowLimit}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SapiV1MarginIsolatedMarginDataData$new()}}
\item \href{#method-toJSON}{\code{SapiV1MarginIsolatedMarginDataData$toJSON()}}
\item \href{#method-fromJSON}{\code{SapiV1MarginIsolatedMarginDataData$fromJSON()}}
\item \href{#method-toJSONString}{\code{SapiV1MarginIsolatedMarginDataData$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SapiV1MarginIsolatedMarginDataData$fromJSONString()}}
\item \href{#method-clone}{\code{SapiV1MarginIsolatedMarginDataData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1MarginIsolatedMarginDataData$new(
  coin = NULL,
  dailyInterest = NULL,
  borrowLimit = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1MarginIsolatedMarginDataData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1MarginIsolatedMarginDataData$fromJSON(
  SapiV1MarginIsolatedMarginDataDataJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1MarginIsolatedMarginDataData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1MarginIsolatedMarginDataData$fromJSONString(
  SapiV1MarginIsolatedMarginDataDataJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SapiV1MarginIsolatedMarginDataData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
