% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings_api.R
\docType{class}
\name{SavingsApi}
\alias{SavingsApi}
\title{Savings operations}
\format{
An \code{R6Class} generator object
}
\description{
binanceRapi.Savings
}
\section{Methods}{

\describe{
\strong{ SapiV1LendingCustomizedFixedPurchasePost } \emph{ Purchase Fixed/Activity Project (USER_DATA) }
Weight(IP): 1

\itemize{
\item \emph{ @param } project.id character
\item \emph{ @param } lot character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20087} \cr


\item status code : 200 | Generated Purchase Id

\item return type : InlineResponse20087 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingDailyProductListGet } \emph{ Get Flexible Product List (USER_DATA) }
Weight(IP): 1

\itemize{
\item \emph{ @param } status Enum < [ALL, SUBSCRIBABLE, UNSUBSCRIBABLE] >
\item \emph{ @param } featured Enum < [ALL, TRUE] >
\item \emph{ @param } current integer
\item \emph{ @param } size integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_81} ) \cr


\item status code : 200 | List of flexible products

\item return type : array[InlineResponse20081] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingDailyPurchasePost } \emph{ Purchase Flexible Product (USER_DATA) }
Weight(IP): 1

\itemize{
\item \emph{ @param } product.id character
\item \emph{ @param } amount numeric
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20083} \cr


\item status code : 200 | Generated Purchase Id

\item return type : InlineResponse20083 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingDailyRedeemPost } \emph{ Redeem Flexible Product (USER_DATA) }
Weight(IP): 1

\itemize{
\item \emph{ @param } product.id character
\item \emph{ @param } amount numeric
\item \emph{ @param } type Enum < [FAST, NORMAL] >
\item \emph{ @param } recv.window integer


\item status code : 200 | OK

\item return type : object 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingDailyTokenPositionGet } \emph{ Get Flexible Product Position (USER_DATA) }
Weight(IP): 1

\itemize{
\item \emph{ @param } asset character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_85} ) \cr


\item status code : 200 | List of flexible product positions

\item return type : array[InlineResponse20085] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingDailyUserLeftQuotaGet } \emph{ Get Left Daily Purchase Quota of Flexible Product (USER_DATA) }
Weight(IP): 1

\itemize{
\item \emph{ @param } product.id character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20082} \cr


\item status code : 200 | Daily purchase quote of flexible product left

\item return type : InlineResponse20082 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingDailyUserRedemptionQuotaGet } \emph{ Get Left Daily Redemption Quota of Flexible Product (USER_DATA) }
Weight(IP): 1

\itemize{
\item \emph{ @param } product.id character
\item \emph{ @param } type Enum < [FAST, NORMAL] >
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20084} \cr


\item status code : 200 | Daily redemption quota of flexible product left

\item return type : InlineResponse20084 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingPositionChangedPost } \emph{ Change Fixed/Activity Position to Daily Position (USER_DATA) }
- PositionId is mandatory parameter for fixed position.  Weight(IP): 1

\itemize{
\item \emph{ @param } project.id character
\item \emph{ @param } lot character
\item \emph{ @param } position.id character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20091} \cr


\item status code : 200 | Purchase information

\item return type : InlineResponse20091 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingProjectListGet } \emph{ Get Fixed/Activity Project List(USER_DATA) }
Weight(IP): 1

\itemize{
\item \emph{ @param } type Enum < [ACTIVITY, CUSTOMIZED_FIXED] >
\item \emph{ @param } status Enum < [ALL, SUBSCRIBABLE, UNSUBSCRIBABLE] >
\item \emph{ @param } is.sort.asc character
\item \emph{ @param } sort.by Enum < [START_TIME, LOT_SIZ, LOT_SIZE, INTEREST_RATE, DURATION] >
\item \emph{ @param } asset character
\item \emph{ @param } current integer
\item \emph{ @param } size integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_86} ) \cr


\item status code : 200 | List of fixed projects

\item return type : array[InlineResponse20086] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingProjectPositionListGet } \emph{ Get Fixed/Activity Project Position (USER_DATA) }
Weight(IP): 1

\itemize{
\item \emph{ @param } asset character
\item \emph{ @param } project.id character
\item \emph{ @param } status Enum < [ALL, SUBSCRIBABLE, UNSUBSCRIBABLE] >
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_88} ) \cr


\item status code : 200 | List of fixed project positions

\item return type : array[InlineResponse20088] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingUnionAccountGet } \emph{ Lending Account (USER_DATA) }
Weight(IP): 1

\itemize{
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20089} \cr


\item status code : 200 | Lending account

\item return type : InlineResponse20089 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingUnionInterestHistoryGet } \emph{ Get Interest History (USER_DATA) }
- The time between startTime and endTime cannot be longer than 30 days. - If startTime and endTime are both not sent, then the last 30 days&#39; data will be returned.  Weight(IP): 1

\itemize{
\item \emph{ @param } lending.type Enum < [DAILY, ACTIVITY, CUSTOMIZED_FIXED] >
\item \emph{ @param } asset character
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } current integer
\item \emph{ @param } size integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_90} ) \cr


\item status code : 200 | History of interest

\item return type : array[InlineResponse20090] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingUnionPurchaseRecordGet } \emph{ Get Purchase Record (USER_DATA) }
- The time between startTime and endTime cannot be longer than 30 days. - If startTime and endTime are both not sent, then the last 30 days&#39; data will be returned.  Weigh(IP): 1

\itemize{
\item \emph{ @param } lending.type Enum < [DAILY, ACTIVITY, CUSTOMIZED_FIXED] >
\item \emph{ @param } asset character
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } current integer
\item \emph{ @param } size integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{OneOfsavingsFlexiblePurchaseRecordsavingsFixedActivityPurchaseRecord} \cr


\item status code : 200 | List of flexible or Fixed/Activity products

\item return type : OneOfsavingsFlexiblePurchaseRecordsavingsFixedActivityPurchaseRecord 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1LendingUnionRedemptionRecordGet } \emph{ Get Redemption Record (USER_DATA) }
- The time between startTime and endTime cannot be longer than 30 days. - If startTime and endTime are both not sent, then the last 30 days&#39; data will be returned.  Weight(IP): 1

\itemize{
\item \emph{ @param } lending.type Enum < [DAILY, ACTIVITY, CUSTOMIZED_FIXED] >
\item \emph{ @param } asset character
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } current integer
\item \emph{ @param } size integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{OneOfsavingsFlexibleRedemptionRecordsavingsFixedActivityRedemptionRecord} \cr


\item status code : 200 | History of Flexible or Fixed/Activity Redemptions

\item return type : OneOfsavingsFlexibleRedemptionRecordsavingsFixedActivityRedemptionRecord 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  SapiV1LendingCustomizedFixedPurchasePost  ####################

library(binanceRapi)
var.project.id <- 'project.id_example' # character | 
var.lot <- 'lot_example' # character | 
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Purchase Fixed/Activity Project (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingCustomizedFixedPurchasePost(var.project.id, var.lot, recv.window=var.recv.window)


####################  SapiV1LendingDailyProductListGet  ####################

library(binanceRapi)
var.status <- 'status_example' # character | \"ALL\", \"SUBSCRIBABLE\", \"UNSUBSCRIBABLE\"; Default: 'ALL'
var.featured <- 'featured_example' # character | \"ALL\", \"TRUE\"; Default: \"ALL\"
var.current <- 1 # integer | Current querying page. Start from 1. Default:1
var.size <- 100 # integer | Default:10 Max:100
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get Flexible Product List (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingDailyProductListGet(var.status, featured=var.featured, current=var.current, size=var.size, recv.window=var.recv.window)


####################  SapiV1LendingDailyPurchasePost  ####################

library(binanceRapi)
var.product.id <- 'product.id_example' # character | 
var.amount <- 1.01 # numeric | 
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Purchase Flexible Product (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingDailyPurchasePost(var.product.id, var.amount, recv.window=var.recv.window)


####################  SapiV1LendingDailyRedeemPost  ####################

library(binanceRapi)
var.product.id <- 'product.id_example' # character | 
var.amount <- 1.01 # numeric | 
var.type <- 'type_example' # character | \"FAST\", \"NORMAL\"
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Redeem Flexible Product (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingDailyRedeemPost(var.product.id, var.amount, var.type, recv.window=var.recv.window)


####################  SapiV1LendingDailyTokenPositionGet  ####################

library(binanceRapi)
var.asset <- 'BTC' # character | 
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get Flexible Product Position (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingDailyTokenPositionGet(var.asset, recv.window=var.recv.window)


####################  SapiV1LendingDailyUserLeftQuotaGet  ####################

library(binanceRapi)
var.product.id <- 'product.id_example' # character | 
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get Left Daily Purchase Quota of Flexible Product (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingDailyUserLeftQuotaGet(var.product.id, recv.window=var.recv.window)


####################  SapiV1LendingDailyUserRedemptionQuotaGet  ####################

library(binanceRapi)
var.product.id <- 'product.id_example' # character | 
var.type <- 'type_example' # character | \"FAST\", \"NORMAL\"
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get Left Daily Redemption Quota of Flexible Product (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingDailyUserRedemptionQuotaGet(var.product.id, var.type, recv.window=var.recv.window)


####################  SapiV1LendingPositionChangedPost  ####################

library(binanceRapi)
var.project.id <- 'project.id_example' # character | 
var.lot <- 'lot_example' # character | 
var.position.id <- 'position.id_example' # character | 
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Change Fixed/Activity Position to Daily Position (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingPositionChangedPost(var.project.id, var.lot, position.id=var.position.id, recv.window=var.recv.window)


####################  SapiV1LendingProjectListGet  ####################

library(binanceRapi)
var.type <- 'type_example' # character | \"ACTIVITY\", \"CUSTOMIZED_FIXED\"
var.status <- 'status_example' # character | \"ALL\", \"SUBSCRIBABLE\", \"UNSUBSCRIBABLE\"; Default: 'ALL'
var.is.sort.asc <- 'is.sort.asc_example' # character | default \"true\"
var.sort.by <- 'sort.by_example' # character | \"START_TIME\", \"LOT_SIZE\", \"INTEREST_RATE\", \"DURATION\"; default \"START_TIME
var.asset <- 'BNB' # character | 
var.current <- 1 # integer | Current querying page. Start from 1. Default:1
var.size <- 100 # integer | Default:10 Max:100
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get Fixed/Activity Project List(USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingProjectListGet(var.type, var.status, var.is.sort.asc, var.sort.by, asset=var.asset, current=var.current, size=var.size, recv.window=var.recv.window)


####################  SapiV1LendingProjectPositionListGet  ####################

library(binanceRapi)
var.asset <- 'BTC' # character | 
var.project.id <- 'project.id_example' # character | 
var.status <- 'status_example' # character | \"ALL\", \"SUBSCRIBABLE\", \"UNSUBSCRIBABLE\"; Default: 'ALL'
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get Fixed/Activity Project Position (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingProjectPositionListGet(var.asset, var.project.id, var.status, recv.window=var.recv.window)


####################  SapiV1LendingUnionAccountGet  ####################

library(binanceRapi)
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Lending Account (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingUnionAccountGet(recv.window=var.recv.window)


####################  SapiV1LendingUnionInterestHistoryGet  ####################

library(binanceRapi)
var.lending.type <- 'lending.type_example' # character | \"DAILY\" for flexible, \"ACTIVITY\" for activity, \"CUSTOMIZED_FIXED\" for fixed
var.asset <- 'BNB' # character | 
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.current <- 1 # integer | Current querying page. Start from 1. Default:1
var.size <- 100 # integer | Default:10 Max:100
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get Interest History (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingUnionInterestHistoryGet(var.lending.type, asset=var.asset, start.time=var.start.time, end.time=var.end.time, current=var.current, size=var.size, recv.window=var.recv.window)


####################  SapiV1LendingUnionPurchaseRecordGet  ####################

library(binanceRapi)
var.lending.type <- 'lending.type_example' # character | \"DAILY\" for flexible, \"ACTIVITY\" for activity, \"CUSTOMIZED_FIXED\" for fixed
var.asset <- 'BNB' # character | 
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.current <- 1 # integer | Current querying page. Start from 1. Default:1
var.size <- 100 # integer | Default:10 Max:100
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get Purchase Record (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingUnionPurchaseRecordGet(var.lending.type, asset=var.asset, start.time=var.start.time, end.time=var.end.time, current=var.current, size=var.size, recv.window=var.recv.window)


####################  SapiV1LendingUnionRedemptionRecordGet  ####################

library(binanceRapi)
var.lending.type <- 'lending.type_example' # character | \"DAILY\" for flexible, \"ACTIVITY\" for activity, \"CUSTOMIZED_FIXED\" for fixed
var.asset <- 'BNB' # character | 
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.current <- 1 # integer | Current querying page. Start from 1. Default:1
var.size <- 100 # integer | Default:10 Max:100
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Get Redemption Record (USER_DATA)
api.instance <- SavingsApi$new()

result <- api.instance$SapiV1LendingUnionRedemptionRecordGet(var.lending.type, asset=var.asset, start.time=var.start.time, end.time=var.end.time, current=var.current, size=var.size, recv.window=var.recv.window)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SavingsApi$new()}}
\item \href{#method-SapiV1LendingCustomizedFixedPurchasePost}{\code{SavingsApi$SapiV1LendingCustomizedFixedPurchasePost()}}
\item \href{#method-SapiV1LendingCustomizedFixedPurchasePostWithHttpInfo}{\code{SavingsApi$SapiV1LendingCustomizedFixedPurchasePostWithHttpInfo()}}
\item \href{#method-SapiV1LendingDailyProductListGet}{\code{SavingsApi$SapiV1LendingDailyProductListGet()}}
\item \href{#method-SapiV1LendingDailyProductListGetWithHttpInfo}{\code{SavingsApi$SapiV1LendingDailyProductListGetWithHttpInfo()}}
\item \href{#method-SapiV1LendingDailyPurchasePost}{\code{SavingsApi$SapiV1LendingDailyPurchasePost()}}
\item \href{#method-SapiV1LendingDailyPurchasePostWithHttpInfo}{\code{SavingsApi$SapiV1LendingDailyPurchasePostWithHttpInfo()}}
\item \href{#method-SapiV1LendingDailyRedeemPost}{\code{SavingsApi$SapiV1LendingDailyRedeemPost()}}
\item \href{#method-SapiV1LendingDailyRedeemPostWithHttpInfo}{\code{SavingsApi$SapiV1LendingDailyRedeemPostWithHttpInfo()}}
\item \href{#method-SapiV1LendingDailyTokenPositionGet}{\code{SavingsApi$SapiV1LendingDailyTokenPositionGet()}}
\item \href{#method-SapiV1LendingDailyTokenPositionGetWithHttpInfo}{\code{SavingsApi$SapiV1LendingDailyTokenPositionGetWithHttpInfo()}}
\item \href{#method-SapiV1LendingDailyUserLeftQuotaGet}{\code{SavingsApi$SapiV1LendingDailyUserLeftQuotaGet()}}
\item \href{#method-SapiV1LendingDailyUserLeftQuotaGetWithHttpInfo}{\code{SavingsApi$SapiV1LendingDailyUserLeftQuotaGetWithHttpInfo()}}
\item \href{#method-SapiV1LendingDailyUserRedemptionQuotaGet}{\code{SavingsApi$SapiV1LendingDailyUserRedemptionQuotaGet()}}
\item \href{#method-SapiV1LendingDailyUserRedemptionQuotaGetWithHttpInfo}{\code{SavingsApi$SapiV1LendingDailyUserRedemptionQuotaGetWithHttpInfo()}}
\item \href{#method-SapiV1LendingPositionChangedPost}{\code{SavingsApi$SapiV1LendingPositionChangedPost()}}
\item \href{#method-SapiV1LendingPositionChangedPostWithHttpInfo}{\code{SavingsApi$SapiV1LendingPositionChangedPostWithHttpInfo()}}
\item \href{#method-SapiV1LendingProjectListGet}{\code{SavingsApi$SapiV1LendingProjectListGet()}}
\item \href{#method-SapiV1LendingProjectListGetWithHttpInfo}{\code{SavingsApi$SapiV1LendingProjectListGetWithHttpInfo()}}
\item \href{#method-SapiV1LendingProjectPositionListGet}{\code{SavingsApi$SapiV1LendingProjectPositionListGet()}}
\item \href{#method-SapiV1LendingProjectPositionListGetWithHttpInfo}{\code{SavingsApi$SapiV1LendingProjectPositionListGetWithHttpInfo()}}
\item \href{#method-SapiV1LendingUnionAccountGet}{\code{SavingsApi$SapiV1LendingUnionAccountGet()}}
\item \href{#method-SapiV1LendingUnionAccountGetWithHttpInfo}{\code{SavingsApi$SapiV1LendingUnionAccountGetWithHttpInfo()}}
\item \href{#method-SapiV1LendingUnionInterestHistoryGet}{\code{SavingsApi$SapiV1LendingUnionInterestHistoryGet()}}
\item \href{#method-SapiV1LendingUnionInterestHistoryGetWithHttpInfo}{\code{SavingsApi$SapiV1LendingUnionInterestHistoryGetWithHttpInfo()}}
\item \href{#method-SapiV1LendingUnionPurchaseRecordGet}{\code{SavingsApi$SapiV1LendingUnionPurchaseRecordGet()}}
\item \href{#method-SapiV1LendingUnionPurchaseRecordGetWithHttpInfo}{\code{SavingsApi$SapiV1LendingUnionPurchaseRecordGetWithHttpInfo()}}
\item \href{#method-SapiV1LendingUnionRedemptionRecordGet}{\code{SavingsApi$SapiV1LendingUnionRedemptionRecordGet()}}
\item \href{#method-SapiV1LendingUnionRedemptionRecordGetWithHttpInfo}{\code{SavingsApi$SapiV1LendingUnionRedemptionRecordGetWithHttpInfo()}}
\item \href{#method-clone}{\code{SavingsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingCustomizedFixedPurchasePost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingCustomizedFixedPurchasePost}{}}}
\subsection{Method \code{SapiV1LendingCustomizedFixedPurchasePost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingCustomizedFixedPurchasePost(
  project.id,
  lot,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingCustomizedFixedPurchasePostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingCustomizedFixedPurchasePostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingCustomizedFixedPurchasePostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingCustomizedFixedPurchasePostWithHttpInfo(
  project.id,
  lot,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyProductListGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyProductListGet}{}}}
\subsection{Method \code{SapiV1LendingDailyProductListGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyProductListGet(
  status,
  featured = NULL,
  current = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyProductListGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyProductListGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingDailyProductListGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyProductListGetWithHttpInfo(
  status,
  featured = NULL,
  current = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyPurchasePost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyPurchasePost}{}}}
\subsection{Method \code{SapiV1LendingDailyPurchasePost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyPurchasePost(
  product.id,
  amount,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyPurchasePostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyPurchasePostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingDailyPurchasePostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyPurchasePostWithHttpInfo(
  product.id,
  amount,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyRedeemPost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyRedeemPost}{}}}
\subsection{Method \code{SapiV1LendingDailyRedeemPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyRedeemPost(
  product.id,
  amount,
  type,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyRedeemPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyRedeemPostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingDailyRedeemPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyRedeemPostWithHttpInfo(
  product.id,
  amount,
  type,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyTokenPositionGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyTokenPositionGet}{}}}
\subsection{Method \code{SapiV1LendingDailyTokenPositionGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyTokenPositionGet(asset, recv.window = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyTokenPositionGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyTokenPositionGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingDailyTokenPositionGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyTokenPositionGetWithHttpInfo(
  asset,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyUserLeftQuotaGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyUserLeftQuotaGet}{}}}
\subsection{Method \code{SapiV1LendingDailyUserLeftQuotaGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyUserLeftQuotaGet(
  product.id,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyUserLeftQuotaGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyUserLeftQuotaGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingDailyUserLeftQuotaGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyUserLeftQuotaGetWithHttpInfo(
  product.id,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyUserRedemptionQuotaGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyUserRedemptionQuotaGet}{}}}
\subsection{Method \code{SapiV1LendingDailyUserRedemptionQuotaGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyUserRedemptionQuotaGet(
  product.id,
  type,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingDailyUserRedemptionQuotaGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingDailyUserRedemptionQuotaGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingDailyUserRedemptionQuotaGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingDailyUserRedemptionQuotaGetWithHttpInfo(
  product.id,
  type,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingPositionChangedPost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingPositionChangedPost}{}}}
\subsection{Method \code{SapiV1LendingPositionChangedPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingPositionChangedPost(
  project.id,
  lot,
  position.id = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingPositionChangedPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingPositionChangedPostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingPositionChangedPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingPositionChangedPostWithHttpInfo(
  project.id,
  lot,
  position.id = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingProjectListGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingProjectListGet}{}}}
\subsection{Method \code{SapiV1LendingProjectListGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingProjectListGet(
  type,
  status,
  is.sort.asc,
  sort.by,
  asset = NULL,
  current = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingProjectListGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingProjectListGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingProjectListGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingProjectListGetWithHttpInfo(
  type,
  status,
  is.sort.asc,
  sort.by,
  asset = NULL,
  current = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingProjectPositionListGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingProjectPositionListGet}{}}}
\subsection{Method \code{SapiV1LendingProjectPositionListGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingProjectPositionListGet(
  asset,
  project.id,
  status,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingProjectPositionListGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingProjectPositionListGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingProjectPositionListGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingProjectPositionListGetWithHttpInfo(
  asset,
  project.id,
  status,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingUnionAccountGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingUnionAccountGet}{}}}
\subsection{Method \code{SapiV1LendingUnionAccountGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingUnionAccountGet(recv.window = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingUnionAccountGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingUnionAccountGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingUnionAccountGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingUnionAccountGetWithHttpInfo(recv.window = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingUnionInterestHistoryGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingUnionInterestHistoryGet}{}}}
\subsection{Method \code{SapiV1LendingUnionInterestHistoryGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingUnionInterestHistoryGet(
  lending.type,
  asset = NULL,
  start.time = NULL,
  end.time = NULL,
  current = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingUnionInterestHistoryGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingUnionInterestHistoryGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingUnionInterestHistoryGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingUnionInterestHistoryGetWithHttpInfo(
  lending.type,
  asset = NULL,
  start.time = NULL,
  end.time = NULL,
  current = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingUnionPurchaseRecordGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingUnionPurchaseRecordGet}{}}}
\subsection{Method \code{SapiV1LendingUnionPurchaseRecordGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingUnionPurchaseRecordGet(
  lending.type,
  asset = NULL,
  start.time = NULL,
  end.time = NULL,
  current = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingUnionPurchaseRecordGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingUnionPurchaseRecordGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingUnionPurchaseRecordGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingUnionPurchaseRecordGetWithHttpInfo(
  lending.type,
  asset = NULL,
  start.time = NULL,
  end.time = NULL,
  current = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingUnionRedemptionRecordGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingUnionRedemptionRecordGet}{}}}
\subsection{Method \code{SapiV1LendingUnionRedemptionRecordGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingUnionRedemptionRecordGet(
  lending.type,
  asset = NULL,
  start.time = NULL,
  end.time = NULL,
  current = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1LendingUnionRedemptionRecordGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1LendingUnionRedemptionRecordGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1LendingUnionRedemptionRecordGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$SapiV1LendingUnionRedemptionRecordGetWithHttpInfo(
  lending.type,
  asset = NULL,
  start.time = NULL,
  end.time = NULL,
  current = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
