% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings_fixed_activity_redemption_record.R
\docType{class}
\name{SavingsFixedActivityRedemptionRecord}
\alias{SavingsFixedActivityRedemptionRecord}
\title{SavingsFixedActivityRedemptionRecord}
\format{
An \code{R6Class} generator object
}
\description{
SavingsFixedActivityRedemptionRecord Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SavingsFixedActivityRedemptionRecord$new()}}
\item \href{#method-toJSON}{\code{SavingsFixedActivityRedemptionRecord$toJSON()}}
\item \href{#method-fromJSON}{\code{SavingsFixedActivityRedemptionRecord$fromJSON()}}
\item \href{#method-toJSONString}{\code{SavingsFixedActivityRedemptionRecord$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SavingsFixedActivityRedemptionRecord$fromJSONString()}}
\item \href{#method-clone}{\code{SavingsFixedActivityRedemptionRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFixedActivityRedemptionRecord$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFixedActivityRedemptionRecord$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFixedActivityRedemptionRecord$fromJSON(
  SavingsFixedActivityRedemptionRecordJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFixedActivityRedemptionRecord$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFixedActivityRedemptionRecord$fromJSONString(
  SavingsFixedActivityRedemptionRecordJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFixedActivityRedemptionRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
