% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings_flexible_redemption_record.R
\docType{class}
\name{SavingsFlexibleRedemptionRecord}
\alias{SavingsFlexibleRedemptionRecord}
\title{SavingsFlexibleRedemptionRecord}
\format{
An \code{R6Class} generator object
}
\description{
SavingsFlexibleRedemptionRecord Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SavingsFlexibleRedemptionRecord$new()}}
\item \href{#method-toJSON}{\code{SavingsFlexibleRedemptionRecord$toJSON()}}
\item \href{#method-fromJSON}{\code{SavingsFlexibleRedemptionRecord$fromJSON()}}
\item \href{#method-toJSONString}{\code{SavingsFlexibleRedemptionRecord$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SavingsFlexibleRedemptionRecord$fromJSONString()}}
\item \href{#method-clone}{\code{SavingsFlexibleRedemptionRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFlexibleRedemptionRecord$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFlexibleRedemptionRecord$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFlexibleRedemptionRecord$fromJSON(SavingsFlexibleRedemptionRecordJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFlexibleRedemptionRecord$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFlexibleRedemptionRecord$fromJSONString(
  SavingsFlexibleRedemptionRecordJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsFlexibleRedemptionRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
