% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_futures_data_assets.R
\docType{class}
\name{SnapshotFuturesDataAssets}
\alias{SnapshotFuturesDataAssets}
\title{SnapshotFuturesDataAssets}
\format{
An \code{R6Class} generator object
}
\description{
SnapshotFuturesDataAssets Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{asset}}{character}

\item{\code{marginBalance}}{character}

\item{\code{walletBalance}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SnapshotFuturesDataAssets$new()}}
\item \href{#method-toJSON}{\code{SnapshotFuturesDataAssets$toJSON()}}
\item \href{#method-fromJSON}{\code{SnapshotFuturesDataAssets$fromJSON()}}
\item \href{#method-toJSONString}{\code{SnapshotFuturesDataAssets$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SnapshotFuturesDataAssets$fromJSONString()}}
\item \href{#method-clone}{\code{SnapshotFuturesDataAssets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataAssets$new(asset, marginBalance, walletBalance, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataAssets$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataAssets$fromJSON(SnapshotFuturesDataAssetsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataAssets$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataAssets$fromJSONString(SnapshotFuturesDataAssetsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataAssets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
