% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_futures_data_position.R
\docType{class}
\name{SnapshotFuturesDataPosition}
\alias{SnapshotFuturesDataPosition}
\title{SnapshotFuturesDataPosition}
\format{
An \code{R6Class} generator object
}
\description{
SnapshotFuturesDataPosition Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entryPrice}}{character}

\item{\code{markPrice}}{character}

\item{\code{positionAmt}}{character}

\item{\code{symbol}}{character}

\item{\code{unRealizedProfit}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SnapshotFuturesDataPosition$new()}}
\item \href{#method-toJSON}{\code{SnapshotFuturesDataPosition$toJSON()}}
\item \href{#method-fromJSON}{\code{SnapshotFuturesDataPosition$fromJSON()}}
\item \href{#method-toJSONString}{\code{SnapshotFuturesDataPosition$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SnapshotFuturesDataPosition$fromJSONString()}}
\item \href{#method-clone}{\code{SnapshotFuturesDataPosition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataPosition$new(
  entryPrice,
  markPrice,
  positionAmt,
  symbol,
  unRealizedProfit,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataPosition$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataPosition$fromJSON(SnapshotFuturesDataPositionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataPosition$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataPosition$fromJSONString(SnapshotFuturesDataPositionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesDataPosition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
