% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_futures_snapshot_vos.R
\docType{class}
\name{SnapshotFuturesSnapshotVos}
\alias{SnapshotFuturesSnapshotVos}
\title{SnapshotFuturesSnapshotVos}
\format{
An \code{R6Class} generator object
}
\description{
SnapshotFuturesSnapshotVos Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{SnapshotFuturesData}}

\item{\code{type}}{character}

\item{\code{updateTime}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SnapshotFuturesSnapshotVos$new()}}
\item \href{#method-toJSON}{\code{SnapshotFuturesSnapshotVos$toJSON()}}
\item \href{#method-fromJSON}{\code{SnapshotFuturesSnapshotVos$fromJSON()}}
\item \href{#method-toJSONString}{\code{SnapshotFuturesSnapshotVos$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SnapshotFuturesSnapshotVos$fromJSONString()}}
\item \href{#method-clone}{\code{SnapshotFuturesSnapshotVos$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesSnapshotVos$new(data, type, updateTime, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesSnapshotVos$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesSnapshotVos$fromJSON(SnapshotFuturesSnapshotVosJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesSnapshotVos$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesSnapshotVos$fromJSONString(SnapshotFuturesSnapshotVosJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotFuturesSnapshotVos$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
