% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_margin_data.R
\docType{class}
\name{SnapshotMarginData}
\alias{SnapshotMarginData}
\title{SnapshotMarginData}
\format{
An \code{R6Class} generator object
}
\description{
SnapshotMarginData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{marginLevel}}{character}

\item{\code{totalAssetOfBtc}}{character}

\item{\code{totalLiabilityOfBtc}}{character}

\item{\code{totalNetAssetOfBtc}}{character}

\item{\code{userAssets}}{list( \link{SnapshotMarginDataUserAssets} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SnapshotMarginData$new()}}
\item \href{#method-toJSON}{\code{SnapshotMarginData$toJSON()}}
\item \href{#method-fromJSON}{\code{SnapshotMarginData$fromJSON()}}
\item \href{#method-toJSONString}{\code{SnapshotMarginData$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SnapshotMarginData$fromJSONString()}}
\item \href{#method-clone}{\code{SnapshotMarginData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginData$new(
  marginLevel,
  totalAssetOfBtc,
  totalLiabilityOfBtc,
  totalNetAssetOfBtc,
  userAssets,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginData$fromJSON(SnapshotMarginDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginData$fromJSONString(SnapshotMarginDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
