% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_margin_data_user_assets.R
\docType{class}
\name{SnapshotMarginDataUserAssets}
\alias{SnapshotMarginDataUserAssets}
\title{SnapshotMarginDataUserAssets}
\format{
An \code{R6Class} generator object
}
\description{
SnapshotMarginDataUserAssets Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{asset}}{character}

\item{\code{borrowed}}{character}

\item{\code{free}}{character}

\item{\code{interest}}{character}

\item{\code{locked}}{character}

\item{\code{netAsset}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SnapshotMarginDataUserAssets$new()}}
\item \href{#method-toJSON}{\code{SnapshotMarginDataUserAssets$toJSON()}}
\item \href{#method-fromJSON}{\code{SnapshotMarginDataUserAssets$fromJSON()}}
\item \href{#method-toJSONString}{\code{SnapshotMarginDataUserAssets$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SnapshotMarginDataUserAssets$fromJSONString()}}
\item \href{#method-clone}{\code{SnapshotMarginDataUserAssets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginDataUserAssets$new(
  asset,
  borrowed,
  free,
  interest,
  locked,
  netAsset,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginDataUserAssets$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginDataUserAssets$fromJSON(SnapshotMarginDataUserAssetsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginDataUserAssets$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginDataUserAssets$fromJSONString(SnapshotMarginDataUserAssetsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginDataUserAssets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
