% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_margin_snapshot_vos.R
\docType{class}
\name{SnapshotMarginSnapshotVos}
\alias{SnapshotMarginSnapshotVos}
\title{SnapshotMarginSnapshotVos}
\format{
An \code{R6Class} generator object
}
\description{
SnapshotMarginSnapshotVos Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{SnapshotMarginData}}

\item{\code{type}}{character}

\item{\code{updateTime}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SnapshotMarginSnapshotVos$new()}}
\item \href{#method-toJSON}{\code{SnapshotMarginSnapshotVos$toJSON()}}
\item \href{#method-fromJSON}{\code{SnapshotMarginSnapshotVos$fromJSON()}}
\item \href{#method-toJSONString}{\code{SnapshotMarginSnapshotVos$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SnapshotMarginSnapshotVos$fromJSONString()}}
\item \href{#method-clone}{\code{SnapshotMarginSnapshotVos$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginSnapshotVos$new(data, type, updateTime, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginSnapshotVos$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginSnapshotVos$fromJSON(SnapshotMarginSnapshotVosJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginSnapshotVos$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginSnapshotVos$fromJSONString(SnapshotMarginSnapshotVosJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnapshotMarginSnapshotVos$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
