% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_account_api.R
\docType{class}
\name{SubAccountApi}
\alias{SubAccountApi}
\title{SubAccount operations}
\format{
An \code{R6Class} generator object
}
\description{
binanceRapi.SubAccount
}
\section{Methods}{

\describe{
\strong{ SapiV1CapitalDepositSubAddressGet } \emph{ Sub-account Spot Assets Summary (For Master Account) }
Fetch sub-account deposit address  Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } coin character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } network character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20053} \cr


\item status code : 200 | Deposit address info

\item return type : InlineResponse20053 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1CapitalDepositSubHisrecGet } \emph{ Sub-account Deposit History (For Master Account) }
Fetch sub-account deposit history  Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } coin character
\item \emph{ @param } status integer
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } limit integer
\item \emph{ @param } offset integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_54} ) \cr


\item status code : 200 | Sub-account deposit history

\item return type : array[InlineResponse20054] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1ManagedSubaccountAssetGet } \emph{ Managed sub-account asset details（For Investor Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_69} ) \cr


\item status code : 200 | List of asset details

\item return type : array[InlineResponse20069] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1ManagedSubaccountDepositPost } \emph{ Deposit assets into the managed sub-account（For Investor Master Account） }
Weight(IP): 1

\itemize{
\item \emph{ @param } to.email character
\item \emph{ @param } asset character
\item \emph{ @param } amount numeric
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20068} \cr


\item status code : 200 | Transfer id

\item return type : InlineResponse20068 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1ManagedSubaccountWithdrawPost } \emph{ Withdrawl assets from the managed sub-account（For Investor Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } from.email character
\item \emph{ @param } asset character
\item \emph{ @param } amount numeric
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } transfer.date integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20068} \cr


\item status code : 200 | Transfer id

\item return type : InlineResponse20068 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountBlvtEnablePost } \emph{ Enable Leverage Token for Sub-account (For Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } enable.blvt character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20067} \cr


\item status code : 200 | BLVT status

\item return type : InlineResponse20067 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountFuturesAccountGet } \emph{ Detail on Sub-account&#39;s Futures Account (For Master Account) }
Weight(IP): 10

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20060} \cr


\item status code : 200 | Futures account details

\item return type : InlineResponse20060 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountFuturesAccountSummaryGet } \emph{ Summary of Sub-account&#39;s Futures Account (For Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20061} \cr


\item status code : 200 | Futures account summary

\item return type : InlineResponse20061 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountFuturesEnablePost } \emph{ Enable Futures for Sub-account (For Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20059} \cr


\item status code : 200 | Futures status

\item return type : InlineResponse20059 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountFuturesInternalTransferGet } \emph{ Sub-account Futures Asset Transfer History (For Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } futures.type integer
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } page integer
\item \emph{ @param } limit integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20049} \cr


\item status code : 200 | Sub-account Futures Asset Transfer History

\item return type : InlineResponse20049 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountFuturesInternalTransferPost } \emph{ Sub-account Futures Asset Transfer (For Master Account) }
- Master account can transfer max 2000 times a minute  Weight(IP): 1

\itemize{
\item \emph{ @param } from.email character
\item \emph{ @param } to.email character
\item \emph{ @param } futures.type integer
\item \emph{ @param } asset character
\item \emph{ @param } amount numeric
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20050} \cr


\item status code : 200 | Futures Asset Transfer Info

\item return type : InlineResponse20050 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountFuturesPositionRiskGet } \emph{ Futures Position-Risk of Sub-account (For Master Account) }
Weight(IP): 10

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_62} ) \cr


\item status code : 200 | Futures account summary

\item return type : array[InlineResponse20062] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountFuturesTransferPost } \emph{ Transfer for Sub-account (For Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } asset character
\item \emph{ @param } amount numeric
\item \emph{ @param } type Enum < [1, 2, 3, 4] >
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20063} \cr


\item status code : 200 | Transfer id

\item return type : InlineResponse20063 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountListGet } \emph{ Query Sub-account List (For Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } email character
\item \emph{ @param } is.freeze Enum < [true, false] >
\item \emph{ @param } page integer
\item \emph{ @param } limit integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20047} \cr


\item status code : 200 | List of sub-accounts

\item return type : InlineResponse20047 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountMarginAccountGet } \emph{ Detail on Sub-account&#39;s Margin Account (For Master Account) }
Weight(IP): 10

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20057} \cr


\item status code : 200 | Margin sub-account details

\item return type : InlineResponse20057 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountMarginAccountSummaryGet } \emph{ Summary of Sub-account&#39;s Margin Account (For Master Account) }
Weight(IP): 10

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20058} \cr


\item status code : 200 | Margin sub-account details

\item return type : InlineResponse20058 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountMarginEnablePost } \emph{ Enable Margin for Sub-account (For Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20056} \cr


\item status code : 200 | Margin status

\item return type : InlineResponse20056 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountMarginTransferPost } \emph{ Margin Transfer for Sub-account (For Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } asset character
\item \emph{ @param } amount numeric
\item \emph{ @param } type Enum < [1, 2] >
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20063} \cr


\item status code : 200 | Transfer id

\item return type : InlineResponse20063 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountSpotSummaryGet } \emph{ Sub-account Spot Assets Summary (For Master Account) }
Get BTC valued asset summary of subaccounts.  Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } page integer
\item \emph{ @param } size integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20052} \cr


\item status code : 200 | Summary of Sub-account Spot Assets

\item return type : InlineResponse20052 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountStatusGet } \emph{ Sub-account&#39;s Status on Margin/Futures (For Master Account) }
- If no &#x60;email&#x60; sent, all sub-accounts&#39; information will be returned.  Weight(IP): 10

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } email character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_55} ) \cr


\item status code : 200 | Status on Margin/Futures

\item return type : array[InlineResponse20055] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountSubTransferHistoryGet } \emph{ Sub-account Spot Asset Transfer History (For Master Account) }
- fromEmail and toEmail cannot be sent at the same time. - Return fromEmail equal master account email by default.  Weight(IP): 1

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } from.email character
\item \emph{ @param } to.email character
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } page integer
\item \emph{ @param } limit integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_48} ) \cr


\item status code : 200 | Sub-account Spot Asset Transfer History

\item return type : array[InlineResponse20048] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountTransferSubToMasterPost } \emph{ Transfer to Master (For Sub-account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } asset character
\item \emph{ @param } amount numeric
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20063} \cr


\item status code : 200 | Transfer id

\item return type : InlineResponse20063 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountTransferSubToSubPost } \emph{ Transfer to Sub-account of Same Master (For Sub-account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } to.email character
\item \emph{ @param } asset character
\item \emph{ @param } amount numeric
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20063} \cr


\item status code : 200 | Transfer id

\item return type : InlineResponse20063 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountTransferSubUserHistoryGet } \emph{ Sub-account Transfer History (For Sub-account) }
- If &#x60;type&#x60; is not sent, the records of type 2: transfer out will be returned by default. - If &#x60;startTime&#x60; and &#x60;endTime&#x60; are not sent, the recent 30-day data will be returned.  Weight(IP): 1

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } asset character
\item \emph{ @param } type Enum < [1, 2] >
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } limit integer
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_64} ) \cr


\item status code : 200 | Transfer id

\item return type : array[InlineResponse20064] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountUniversalTransferGet } \emph{ Universal Transfer History (For Master Account) }
- fromEmail and toEmail cannot be sent at the same time. - Return fromEmail equal master account email by default. - Only get the latest history of past 30 days.  Weight(IP): 1

\itemize{
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } from.email character
\item \emph{ @param } to.email character
\item \emph{ @param } start.time integer
\item \emph{ @param } end.time integer
\item \emph{ @param } page integer
\item \emph{ @param } limit character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } list( \link{inline_response_200_65} ) \cr


\item status code : 200 | Transfer History

\item return type : array[InlineResponse20065] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountUniversalTransferPost } \emph{ Universal Transfer (For Master Account) }
- You need to enable \&quot;internal transfer\&quot; option for the api key which requests this endpoint. - Transfer from master account by default if fromEmail is not sent. - Transfer to master account by default if toEmail is not sent. - Transfer between futures accounts is not supported.  Weight(IP): 1

\itemize{
\item \emph{ @param } from.account.type Enum < [SPOT, USDT_FUTURE, COIN_FUTURE] >
\item \emph{ @param } to.account.type Enum < [SPOT, USDT_FUTURE, COIN_FUTURE] >
\item \emph{ @param } asset character
\item \emph{ @param } amount numeric
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } from.email character
\item \emph{ @param } to.email character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20066} \cr


\item status code : 200 | Transfer id

\item return type : InlineResponse20066 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV1SubAccountVirtualSubAccountPost } \emph{ Create a Virtual Sub-account(For Master Account) }
- This request will generate a virtual sub account under your master account. - You need to enable \&quot;trade\&quot; option for the api key which requests this endpoint.  Weight(IP): 1

\itemize{
\item \emph{ @param } sub.account.string character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20046} \cr


\item status code : 200 | Return the created virtual email

\item return type : InlineResponse20046 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV2SubAccountFuturesAccountGet } \emph{ Detail on Sub-account&#39;s Futures Account V2 (For Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } futures.type Enum < [1, 2] >
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{OneOfsubAccountUSDTFuturesDetailssubAccountCOINFuturesDetails} \cr


\item status code : 200 | USDT or COIN Margined Futures Details

\item return type : OneOfsubAccountUSDTFuturesDetailssubAccountCOINFuturesDetails 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV2SubAccountFuturesAccountSummaryGet } \emph{ Summary of Sub-account&#39;s Futures Account V2 (For Master Account) }
Weight(IP): 10

\itemize{
\item \emph{ @param } futures.type Enum < [1, 2] >
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } page integer
\item \emph{ @param } limit character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{OneOfsubAccountUSDTFuturesSummarysubAccountCOINFuturesSummary} \cr


\item status code : 200 | USDT or COIN Margined Futures Summary

\item return type : OneOfsubAccountUSDTFuturesSummarysubAccountCOINFuturesSummary 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV2SubAccountFuturesPositionRiskGet } \emph{ Futures Position-Risk of Sub-account V2 (For Master Account) }
Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } futures.type Enum < [1, 2] >
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{OneOfsubAccountUSDTFuturesPositionRisksubAccountCOINFuturesPositionRisk} \cr


\item status code : 200 | USDT or COIN Margined Futures Position Risk

\item return type : OneOfsubAccountUSDTFuturesPositionRisksubAccountCOINFuturesPositionRisk 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ SapiV3SubAccountAssetsGet } \emph{ Sub-account Assets (For Master Account) }
Fetch sub-account assets  Weight(IP): 1

\itemize{
\item \emph{ @param } email character
\item \emph{ @param } timestamp integer
\item \emph{ @param } signature character
\item \emph{ @param } recv.window integer
\item \emph{ @returnType } \link{InlineResponse20051} \cr


\item status code : 200 | List of assets balances

\item return type : InlineResponse20051 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 401 | Unauthorized Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  SapiV1CapitalDepositSubAddressGet  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.coin <- 'BNB' # character | Coin name
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.network <- 'network_example' # character | 
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Sub-account Spot Assets Summary (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1CapitalDepositSubAddressGet(var.email, var.coin, var.timestamp, var.signature, network=var.network, recv.window=var.recv.window)


####################  SapiV1CapitalDepositSubHisrecGet  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.coin <- 'BNB' # character | Coin name
var.status <- 56 # integer | 0(0:pending,6: credited but cannot withdraw, 1:success)
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.limit <- 56 # integer | 
var.offset <- 56 # integer | 
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Sub-account Deposit History (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1CapitalDepositSubHisrecGet(var.email, var.timestamp, var.signature, coin=var.coin, status=var.status, start.time=var.start.time, end.time=var.end.time, limit=var.limit, offset=var.offset, recv.window=var.recv.window)


####################  SapiV1ManagedSubaccountAssetGet  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Managed sub-account asset details（For Investor Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1ManagedSubaccountAssetGet(var.email, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1ManagedSubaccountDepositPost  ####################

library(binanceRapi)
var.to.email <- 'to.email_example' # character | Recipient email
var.asset <- 'BTC' # character | 
var.amount <- 1.01 # numeric | 
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Deposit assets into the managed sub-account（For Investor Master Account）
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1ManagedSubaccountDepositPost(var.to.email, var.asset, var.amount, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1ManagedSubaccountWithdrawPost  ####################

library(binanceRapi)
var.from.email <- 'from.email_example' # character | Sender email
var.asset <- 'BTC' # character | 
var.amount <- 1.01 # numeric | 
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.transfer.date <- 56 # integer | Withdrawals is automatically occur on the transfer date(UTC0). If a date is not selected, the withdrawal occurs right now
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Withdrawl assets from the managed sub-account（For Investor Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1ManagedSubaccountWithdrawPost(var.from.email, var.asset, var.amount, var.timestamp, var.signature, transfer.date=var.transfer.date, recv.window=var.recv.window)


####################  SapiV1SubAccountBlvtEnablePost  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.enable.blvt <- 'enable.blvt_example' # character | Only true for now
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Enable Leverage Token for Sub-account (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountBlvtEnablePost(var.email, var.enable.blvt, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountFuturesAccountGet  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Detail on Sub-account's Futures Account (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountFuturesAccountGet(var.email, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountFuturesAccountSummaryGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Summary of Sub-account's Futures Account (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountFuturesAccountSummaryGet(var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountFuturesEnablePost  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Enable Futures for Sub-account (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountFuturesEnablePost(var.email, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountFuturesInternalTransferGet  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.futures.type <- 2 # integer | 1:USDT-margined Futures, 2: Coin-margined Futures
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.page <- 1 # integer | Default 1
var.limit <- 56 # integer | Default value: 50, Max value: 500
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Sub-account Futures Asset Transfer History (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountFuturesInternalTransferGet(var.email, var.futures.type, var.timestamp, var.signature, start.time=var.start.time, end.time=var.end.time, page=var.page, limit=var.limit, recv.window=var.recv.window)


####################  SapiV1SubAccountFuturesInternalTransferPost  ####################

library(binanceRapi)
var.from.email <- 'from.email_example' # character | Sender email
var.to.email <- 'to.email_example' # character | Recipient email
var.futures.type <- 2 # integer | 1:USDT-margined Futures,2: Coin-margined Futures
var.asset <- 'BTC' # character | 
var.amount <- 1.01 # numeric | 
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Sub-account Futures Asset Transfer (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountFuturesInternalTransferPost(var.from.email, var.to.email, var.futures.type, var.asset, var.amount, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountFuturesPositionRiskGet  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Futures Position-Risk of Sub-account (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountFuturesPositionRiskGet(var.email, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountFuturesTransferPost  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.asset <- 'BTC' # character | 
var.amount <- 1.01 # numeric | 
var.type <- 56 # integer | 1: transfer from subaccount's spot account to its USDT-margined futures account  2: transfer from subaccount's USDT-margined futures account to its spot account  3: transfer from subaccount's spot account to its COIN-margined futures account  4:transfer from subaccount's COIN-margined futures account to its spot account
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Transfer for Sub-account (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountFuturesTransferPost(var.email, var.asset, var.amount, var.type, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountListGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.email <- 'email_example' # character | Sub-account email
var.is.freeze <- 'is.freeze_example' # character | 
var.page <- 1 # integer | Default 1
var.limit <- 1 # integer | Default 1; max 200
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Query Sub-account List (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountListGet(var.timestamp, var.signature, email=var.email, is.freeze=var.is.freeze, page=var.page, limit=var.limit, recv.window=var.recv.window)


####################  SapiV1SubAccountMarginAccountGet  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Detail on Sub-account's Margin Account (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountMarginAccountGet(var.email, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountMarginAccountSummaryGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Summary of Sub-account's Margin Account (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountMarginAccountSummaryGet(var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountMarginEnablePost  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Enable Margin for Sub-account (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountMarginEnablePost(var.email, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountMarginTransferPost  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.asset <- 'BTC' # character | 
var.amount <- 1.01 # numeric | 
var.type <- 56 # integer | 1: transfer from subaccount's spot account to margin account  2: transfer from subaccount's margin account to its spot account
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Margin Transfer for Sub-account (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountMarginTransferPost(var.email, var.asset, var.amount, var.type, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountSpotSummaryGet  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.page <- 1 # integer | Default 1
var.size <- 56 # integer | Default:10 Max:20
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Sub-account Spot Assets Summary (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountSpotSummaryGet(var.email, var.timestamp, var.signature, page=var.page, size=var.size, recv.window=var.recv.window)


####################  SapiV1SubAccountStatusGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.email <- 'email_example' # character | Sub-account email
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Sub-account's Status on Margin/Futures (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountStatusGet(var.timestamp, var.signature, email=var.email, recv.window=var.recv.window)


####################  SapiV1SubAccountSubTransferHistoryGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.from.email <- 'from.email_example' # character | Sub-account email
var.to.email <- 'to.email_example' # character | Sub-account email
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.page <- 1 # integer | Default 1
var.limit <- 1 # integer | Default 1
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Sub-account Spot Asset Transfer History (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountSubTransferHistoryGet(var.timestamp, var.signature, from.email=var.from.email, to.email=var.to.email, start.time=var.start.time, end.time=var.end.time, page=var.page, limit=var.limit, recv.window=var.recv.window)


####################  SapiV1SubAccountTransferSubToMasterPost  ####################

library(binanceRapi)
var.asset <- 'BTC' # character | 
var.amount <- 1.01 # numeric | 
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Transfer to Master (For Sub-account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountTransferSubToMasterPost(var.asset, var.amount, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountTransferSubToSubPost  ####################

library(binanceRapi)
var.to.email <- 'to.email_example' # character | Recipient email
var.asset <- 'BTC' # character | 
var.amount <- 1.01 # numeric | 
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Transfer to Sub-account of Same Master (For Sub-account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountTransferSubToSubPost(var.to.email, var.asset, var.amount, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV1SubAccountTransferSubUserHistoryGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.asset <- 'BNB' # character | 
var.type <- 56 # integer | 1: transfer in  2:  transfer out
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.limit <- 500 # integer | Default 500; max 1000.
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Sub-account Transfer History (For Sub-account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountTransferSubUserHistoryGet(var.timestamp, var.signature, asset=var.asset, type=var.type, start.time=var.start.time, end.time=var.end.time, limit=var.limit, recv.window=var.recv.window)


####################  SapiV1SubAccountUniversalTransferGet  ####################

library(binanceRapi)
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.from.email <- 'from.email_example' # character | Sub-account email
var.to.email <- 'to.email_example' # character | Sub-account email
var.start.time <- 56 # integer | UTC timestamp in ms
var.end.time <- 56 # integer | UTC timestamp in ms
var.page <- 1 # integer | Default 1
var.limit <- 'limit_example' # character | Default 500, Max 500
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Universal Transfer History (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountUniversalTransferGet(var.timestamp, var.signature, from.email=var.from.email, to.email=var.to.email, start.time=var.start.time, end.time=var.end.time, page=var.page, limit=var.limit, recv.window=var.recv.window)


####################  SapiV1SubAccountUniversalTransferPost  ####################

library(binanceRapi)
var.from.account.type <- 'from.account.type_example' # character | 
var.to.account.type <- 'to.account.type_example' # character | 
var.asset <- 'BTC' # character | 
var.amount <- 1.01 # numeric | 
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.from.email <- 'from.email_example' # character | Sub-account email
var.to.email <- 'to.email_example' # character | Sub-account email
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Universal Transfer (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountUniversalTransferPost(var.from.account.type, var.to.account.type, var.asset, var.amount, var.timestamp, var.signature, from.email=var.from.email, to.email=var.to.email, recv.window=var.recv.window)


####################  SapiV1SubAccountVirtualSubAccountPost  ####################

library(binanceRapi)
var.sub.account.string <- 'sub.account.string_example' # character | Please input a string. We will create a virtual email using that string for you to register
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Create a Virtual Sub-account(For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV1SubAccountVirtualSubAccountPost(var.sub.account.string, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV2SubAccountFuturesAccountGet  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.futures.type <- 56 # integer | 1:USDT Margined Futures  2:COIN Margined Futures
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Detail on Sub-account's Futures Account V2 (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV2SubAccountFuturesAccountGet(var.email, var.futures.type, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV2SubAccountFuturesAccountSummaryGet  ####################

library(binanceRapi)
var.futures.type <- 56 # integer | 1:USDT Margined Futures  2:COIN Margined Futures
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.page <- 1 # integer | Default 1
var.limit <- 'limit_example' # character | Default 10, Max 20
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Summary of Sub-account's Futures Account V2 (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV2SubAccountFuturesAccountSummaryGet(var.futures.type, var.timestamp, var.signature, page=var.page, limit=var.limit, recv.window=var.recv.window)


####################  SapiV2SubAccountFuturesPositionRiskGet  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.futures.type <- 56 # integer | 1:USDT Margined Futures  2:COIN Margined Futures
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Futures Position-Risk of Sub-account V2 (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV2SubAccountFuturesPositionRiskGet(var.email, var.futures.type, var.timestamp, var.signature, recv.window=var.recv.window)


####################  SapiV3SubAccountAssetsGet  ####################

library(binanceRapi)
var.email <- 'email_example' # character | Sub-account email
var.timestamp <- 56 # integer | UTC timestamp in ms
var.signature <- 'signature_example' # character | Signature
var.recv.window <- 5000 # integer | The value cannot be greater than 60000

#Sub-account Assets (For Master Account)
api.instance <- SubAccountApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-MBX-APIKEY'] <- 'TODO_YOUR_API_KEY';

result <- api.instance$SapiV3SubAccountAssetsGet(var.email, var.timestamp, var.signature, recv.window=var.recv.window)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubAccountApi$new()}}
\item \href{#method-SapiV1CapitalDepositSubAddressGet}{\code{SubAccountApi$SapiV1CapitalDepositSubAddressGet()}}
\item \href{#method-SapiV1CapitalDepositSubAddressGetWithHttpInfo}{\code{SubAccountApi$SapiV1CapitalDepositSubAddressGetWithHttpInfo()}}
\item \href{#method-SapiV1CapitalDepositSubHisrecGet}{\code{SubAccountApi$SapiV1CapitalDepositSubHisrecGet()}}
\item \href{#method-SapiV1CapitalDepositSubHisrecGetWithHttpInfo}{\code{SubAccountApi$SapiV1CapitalDepositSubHisrecGetWithHttpInfo()}}
\item \href{#method-SapiV1ManagedSubaccountAssetGet}{\code{SubAccountApi$SapiV1ManagedSubaccountAssetGet()}}
\item \href{#method-SapiV1ManagedSubaccountAssetGetWithHttpInfo}{\code{SubAccountApi$SapiV1ManagedSubaccountAssetGetWithHttpInfo()}}
\item \href{#method-SapiV1ManagedSubaccountDepositPost}{\code{SubAccountApi$SapiV1ManagedSubaccountDepositPost()}}
\item \href{#method-SapiV1ManagedSubaccountDepositPostWithHttpInfo}{\code{SubAccountApi$SapiV1ManagedSubaccountDepositPostWithHttpInfo()}}
\item \href{#method-SapiV1ManagedSubaccountWithdrawPost}{\code{SubAccountApi$SapiV1ManagedSubaccountWithdrawPost()}}
\item \href{#method-SapiV1ManagedSubaccountWithdrawPostWithHttpInfo}{\code{SubAccountApi$SapiV1ManagedSubaccountWithdrawPostWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountBlvtEnablePost}{\code{SubAccountApi$SapiV1SubAccountBlvtEnablePost()}}
\item \href{#method-SapiV1SubAccountBlvtEnablePostWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountBlvtEnablePostWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountFuturesAccountGet}{\code{SubAccountApi$SapiV1SubAccountFuturesAccountGet()}}
\item \href{#method-SapiV1SubAccountFuturesAccountGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountFuturesAccountGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountFuturesAccountSummaryGet}{\code{SubAccountApi$SapiV1SubAccountFuturesAccountSummaryGet()}}
\item \href{#method-SapiV1SubAccountFuturesAccountSummaryGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountFuturesAccountSummaryGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountFuturesEnablePost}{\code{SubAccountApi$SapiV1SubAccountFuturesEnablePost()}}
\item \href{#method-SapiV1SubAccountFuturesEnablePostWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountFuturesEnablePostWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountFuturesInternalTransferGet}{\code{SubAccountApi$SapiV1SubAccountFuturesInternalTransferGet()}}
\item \href{#method-SapiV1SubAccountFuturesInternalTransferGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountFuturesInternalTransferGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountFuturesInternalTransferPost}{\code{SubAccountApi$SapiV1SubAccountFuturesInternalTransferPost()}}
\item \href{#method-SapiV1SubAccountFuturesInternalTransferPostWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountFuturesInternalTransferPostWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountFuturesPositionRiskGet}{\code{SubAccountApi$SapiV1SubAccountFuturesPositionRiskGet()}}
\item \href{#method-SapiV1SubAccountFuturesPositionRiskGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountFuturesPositionRiskGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountFuturesTransferPost}{\code{SubAccountApi$SapiV1SubAccountFuturesTransferPost()}}
\item \href{#method-SapiV1SubAccountFuturesTransferPostWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountFuturesTransferPostWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountListGet}{\code{SubAccountApi$SapiV1SubAccountListGet()}}
\item \href{#method-SapiV1SubAccountListGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountListGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountMarginAccountGet}{\code{SubAccountApi$SapiV1SubAccountMarginAccountGet()}}
\item \href{#method-SapiV1SubAccountMarginAccountGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountMarginAccountGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountMarginAccountSummaryGet}{\code{SubAccountApi$SapiV1SubAccountMarginAccountSummaryGet()}}
\item \href{#method-SapiV1SubAccountMarginAccountSummaryGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountMarginAccountSummaryGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountMarginEnablePost}{\code{SubAccountApi$SapiV1SubAccountMarginEnablePost()}}
\item \href{#method-SapiV1SubAccountMarginEnablePostWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountMarginEnablePostWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountMarginTransferPost}{\code{SubAccountApi$SapiV1SubAccountMarginTransferPost()}}
\item \href{#method-SapiV1SubAccountMarginTransferPostWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountMarginTransferPostWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountSpotSummaryGet}{\code{SubAccountApi$SapiV1SubAccountSpotSummaryGet()}}
\item \href{#method-SapiV1SubAccountSpotSummaryGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountSpotSummaryGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountStatusGet}{\code{SubAccountApi$SapiV1SubAccountStatusGet()}}
\item \href{#method-SapiV1SubAccountStatusGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountStatusGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountSubTransferHistoryGet}{\code{SubAccountApi$SapiV1SubAccountSubTransferHistoryGet()}}
\item \href{#method-SapiV1SubAccountSubTransferHistoryGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountSubTransferHistoryGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountTransferSubToMasterPost}{\code{SubAccountApi$SapiV1SubAccountTransferSubToMasterPost()}}
\item \href{#method-SapiV1SubAccountTransferSubToMasterPostWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountTransferSubToMasterPostWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountTransferSubToSubPost}{\code{SubAccountApi$SapiV1SubAccountTransferSubToSubPost()}}
\item \href{#method-SapiV1SubAccountTransferSubToSubPostWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountTransferSubToSubPostWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountTransferSubUserHistoryGet}{\code{SubAccountApi$SapiV1SubAccountTransferSubUserHistoryGet()}}
\item \href{#method-SapiV1SubAccountTransferSubUserHistoryGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountTransferSubUserHistoryGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountUniversalTransferGet}{\code{SubAccountApi$SapiV1SubAccountUniversalTransferGet()}}
\item \href{#method-SapiV1SubAccountUniversalTransferGetWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountUniversalTransferGetWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountUniversalTransferPost}{\code{SubAccountApi$SapiV1SubAccountUniversalTransferPost()}}
\item \href{#method-SapiV1SubAccountUniversalTransferPostWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountUniversalTransferPostWithHttpInfo()}}
\item \href{#method-SapiV1SubAccountVirtualSubAccountPost}{\code{SubAccountApi$SapiV1SubAccountVirtualSubAccountPost()}}
\item \href{#method-SapiV1SubAccountVirtualSubAccountPostWithHttpInfo}{\code{SubAccountApi$SapiV1SubAccountVirtualSubAccountPostWithHttpInfo()}}
\item \href{#method-SapiV2SubAccountFuturesAccountGet}{\code{SubAccountApi$SapiV2SubAccountFuturesAccountGet()}}
\item \href{#method-SapiV2SubAccountFuturesAccountGetWithHttpInfo}{\code{SubAccountApi$SapiV2SubAccountFuturesAccountGetWithHttpInfo()}}
\item \href{#method-SapiV2SubAccountFuturesAccountSummaryGet}{\code{SubAccountApi$SapiV2SubAccountFuturesAccountSummaryGet()}}
\item \href{#method-SapiV2SubAccountFuturesAccountSummaryGetWithHttpInfo}{\code{SubAccountApi$SapiV2SubAccountFuturesAccountSummaryGetWithHttpInfo()}}
\item \href{#method-SapiV2SubAccountFuturesPositionRiskGet}{\code{SubAccountApi$SapiV2SubAccountFuturesPositionRiskGet()}}
\item \href{#method-SapiV2SubAccountFuturesPositionRiskGetWithHttpInfo}{\code{SubAccountApi$SapiV2SubAccountFuturesPositionRiskGetWithHttpInfo()}}
\item \href{#method-SapiV3SubAccountAssetsGet}{\code{SubAccountApi$SapiV3SubAccountAssetsGet()}}
\item \href{#method-SapiV3SubAccountAssetsGetWithHttpInfo}{\code{SubAccountApi$SapiV3SubAccountAssetsGetWithHttpInfo()}}
\item \href{#method-clone}{\code{SubAccountApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1CapitalDepositSubAddressGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1CapitalDepositSubAddressGet}{}}}
\subsection{Method \code{SapiV1CapitalDepositSubAddressGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1CapitalDepositSubAddressGet(
  email,
  coin,
  timestamp,
  signature,
  network = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1CapitalDepositSubAddressGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1CapitalDepositSubAddressGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1CapitalDepositSubAddressGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1CapitalDepositSubAddressGetWithHttpInfo(
  email,
  coin,
  timestamp,
  signature,
  network = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1CapitalDepositSubHisrecGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1CapitalDepositSubHisrecGet}{}}}
\subsection{Method \code{SapiV1CapitalDepositSubHisrecGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1CapitalDepositSubHisrecGet(
  email,
  timestamp,
  signature,
  coin = NULL,
  status = NULL,
  start.time = NULL,
  end.time = NULL,
  limit = NULL,
  offset = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1CapitalDepositSubHisrecGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1CapitalDepositSubHisrecGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1CapitalDepositSubHisrecGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1CapitalDepositSubHisrecGetWithHttpInfo(
  email,
  timestamp,
  signature,
  coin = NULL,
  status = NULL,
  start.time = NULL,
  end.time = NULL,
  limit = NULL,
  offset = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1ManagedSubaccountAssetGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1ManagedSubaccountAssetGet}{}}}
\subsection{Method \code{SapiV1ManagedSubaccountAssetGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1ManagedSubaccountAssetGet(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1ManagedSubaccountAssetGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1ManagedSubaccountAssetGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1ManagedSubaccountAssetGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1ManagedSubaccountAssetGetWithHttpInfo(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1ManagedSubaccountDepositPost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1ManagedSubaccountDepositPost}{}}}
\subsection{Method \code{SapiV1ManagedSubaccountDepositPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1ManagedSubaccountDepositPost(
  to.email,
  asset,
  amount,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1ManagedSubaccountDepositPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1ManagedSubaccountDepositPostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1ManagedSubaccountDepositPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1ManagedSubaccountDepositPostWithHttpInfo(
  to.email,
  asset,
  amount,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1ManagedSubaccountWithdrawPost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1ManagedSubaccountWithdrawPost}{}}}
\subsection{Method \code{SapiV1ManagedSubaccountWithdrawPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1ManagedSubaccountWithdrawPost(
  from.email,
  asset,
  amount,
  timestamp,
  signature,
  transfer.date = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1ManagedSubaccountWithdrawPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1ManagedSubaccountWithdrawPostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1ManagedSubaccountWithdrawPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1ManagedSubaccountWithdrawPostWithHttpInfo(
  from.email,
  asset,
  amount,
  timestamp,
  signature,
  transfer.date = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountBlvtEnablePost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountBlvtEnablePost}{}}}
\subsection{Method \code{SapiV1SubAccountBlvtEnablePost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountBlvtEnablePost(
  email,
  enable.blvt,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountBlvtEnablePostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountBlvtEnablePostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountBlvtEnablePostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountBlvtEnablePostWithHttpInfo(
  email,
  enable.blvt,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesAccountGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesAccountGet}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesAccountGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesAccountGet(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesAccountGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesAccountGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesAccountGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesAccountGetWithHttpInfo(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesAccountSummaryGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesAccountSummaryGet}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesAccountSummaryGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesAccountSummaryGet(
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesAccountSummaryGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesAccountSummaryGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesAccountSummaryGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesAccountSummaryGetWithHttpInfo(
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesEnablePost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesEnablePost}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesEnablePost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesEnablePost(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesEnablePostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesEnablePostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesEnablePostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesEnablePostWithHttpInfo(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesInternalTransferGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesInternalTransferGet}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesInternalTransferGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesInternalTransferGet(
  email,
  futures.type,
  timestamp,
  signature,
  start.time = NULL,
  end.time = NULL,
  page = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesInternalTransferGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesInternalTransferGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesInternalTransferGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesInternalTransferGetWithHttpInfo(
  email,
  futures.type,
  timestamp,
  signature,
  start.time = NULL,
  end.time = NULL,
  page = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesInternalTransferPost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesInternalTransferPost}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesInternalTransferPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesInternalTransferPost(
  from.email,
  to.email,
  futures.type,
  asset,
  amount,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesInternalTransferPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesInternalTransferPostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesInternalTransferPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesInternalTransferPostWithHttpInfo(
  from.email,
  to.email,
  futures.type,
  asset,
  amount,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesPositionRiskGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesPositionRiskGet}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesPositionRiskGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesPositionRiskGet(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesPositionRiskGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesPositionRiskGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesPositionRiskGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesPositionRiskGetWithHttpInfo(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesTransferPost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesTransferPost}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesTransferPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesTransferPost(
  email,
  asset,
  amount,
  type,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountFuturesTransferPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountFuturesTransferPostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountFuturesTransferPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountFuturesTransferPostWithHttpInfo(
  email,
  asset,
  amount,
  type,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountListGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountListGet}{}}}
\subsection{Method \code{SapiV1SubAccountListGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountListGet(
  timestamp,
  signature,
  email = NULL,
  is.freeze = NULL,
  page = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountListGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountListGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountListGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountListGetWithHttpInfo(
  timestamp,
  signature,
  email = NULL,
  is.freeze = NULL,
  page = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountMarginAccountGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountMarginAccountGet}{}}}
\subsection{Method \code{SapiV1SubAccountMarginAccountGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountMarginAccountGet(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountMarginAccountGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountMarginAccountGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountMarginAccountGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountMarginAccountGetWithHttpInfo(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountMarginAccountSummaryGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountMarginAccountSummaryGet}{}}}
\subsection{Method \code{SapiV1SubAccountMarginAccountSummaryGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountMarginAccountSummaryGet(
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountMarginAccountSummaryGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountMarginAccountSummaryGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountMarginAccountSummaryGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountMarginAccountSummaryGetWithHttpInfo(
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountMarginEnablePost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountMarginEnablePost}{}}}
\subsection{Method \code{SapiV1SubAccountMarginEnablePost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountMarginEnablePost(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountMarginEnablePostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountMarginEnablePostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountMarginEnablePostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountMarginEnablePostWithHttpInfo(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountMarginTransferPost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountMarginTransferPost}{}}}
\subsection{Method \code{SapiV1SubAccountMarginTransferPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountMarginTransferPost(
  email,
  asset,
  amount,
  type,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountMarginTransferPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountMarginTransferPostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountMarginTransferPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountMarginTransferPostWithHttpInfo(
  email,
  asset,
  amount,
  type,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountSpotSummaryGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountSpotSummaryGet}{}}}
\subsection{Method \code{SapiV1SubAccountSpotSummaryGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountSpotSummaryGet(
  email,
  timestamp,
  signature,
  page = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountSpotSummaryGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountSpotSummaryGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountSpotSummaryGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountSpotSummaryGetWithHttpInfo(
  email,
  timestamp,
  signature,
  page = NULL,
  size = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountStatusGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountStatusGet}{}}}
\subsection{Method \code{SapiV1SubAccountStatusGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountStatusGet(
  timestamp,
  signature,
  email = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountStatusGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountStatusGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountStatusGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountStatusGetWithHttpInfo(
  timestamp,
  signature,
  email = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountSubTransferHistoryGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountSubTransferHistoryGet}{}}}
\subsection{Method \code{SapiV1SubAccountSubTransferHistoryGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountSubTransferHistoryGet(
  timestamp,
  signature,
  from.email = NULL,
  to.email = NULL,
  start.time = NULL,
  end.time = NULL,
  page = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountSubTransferHistoryGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountSubTransferHistoryGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountSubTransferHistoryGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountSubTransferHistoryGetWithHttpInfo(
  timestamp,
  signature,
  from.email = NULL,
  to.email = NULL,
  start.time = NULL,
  end.time = NULL,
  page = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountTransferSubToMasterPost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountTransferSubToMasterPost}{}}}
\subsection{Method \code{SapiV1SubAccountTransferSubToMasterPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountTransferSubToMasterPost(
  asset,
  amount,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountTransferSubToMasterPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountTransferSubToMasterPostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountTransferSubToMasterPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountTransferSubToMasterPostWithHttpInfo(
  asset,
  amount,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountTransferSubToSubPost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountTransferSubToSubPost}{}}}
\subsection{Method \code{SapiV1SubAccountTransferSubToSubPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountTransferSubToSubPost(
  to.email,
  asset,
  amount,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountTransferSubToSubPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountTransferSubToSubPostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountTransferSubToSubPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountTransferSubToSubPostWithHttpInfo(
  to.email,
  asset,
  amount,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountTransferSubUserHistoryGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountTransferSubUserHistoryGet}{}}}
\subsection{Method \code{SapiV1SubAccountTransferSubUserHistoryGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountTransferSubUserHistoryGet(
  timestamp,
  signature,
  asset = NULL,
  type = NULL,
  start.time = NULL,
  end.time = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountTransferSubUserHistoryGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountTransferSubUserHistoryGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountTransferSubUserHistoryGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountTransferSubUserHistoryGetWithHttpInfo(
  timestamp,
  signature,
  asset = NULL,
  type = NULL,
  start.time = NULL,
  end.time = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountUniversalTransferGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountUniversalTransferGet}{}}}
\subsection{Method \code{SapiV1SubAccountUniversalTransferGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountUniversalTransferGet(
  timestamp,
  signature,
  from.email = NULL,
  to.email = NULL,
  start.time = NULL,
  end.time = NULL,
  page = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountUniversalTransferGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountUniversalTransferGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountUniversalTransferGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountUniversalTransferGetWithHttpInfo(
  timestamp,
  signature,
  from.email = NULL,
  to.email = NULL,
  start.time = NULL,
  end.time = NULL,
  page = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountUniversalTransferPost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountUniversalTransferPost}{}}}
\subsection{Method \code{SapiV1SubAccountUniversalTransferPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountUniversalTransferPost(
  from.account.type,
  to.account.type,
  asset,
  amount,
  timestamp,
  signature,
  from.email = NULL,
  to.email = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountUniversalTransferPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountUniversalTransferPostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountUniversalTransferPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountUniversalTransferPostWithHttpInfo(
  from.account.type,
  to.account.type,
  asset,
  amount,
  timestamp,
  signature,
  from.email = NULL,
  to.email = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountVirtualSubAccountPost"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountVirtualSubAccountPost}{}}}
\subsection{Method \code{SapiV1SubAccountVirtualSubAccountPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountVirtualSubAccountPost(
  sub.account.string,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV1SubAccountVirtualSubAccountPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV1SubAccountVirtualSubAccountPostWithHttpInfo}{}}}
\subsection{Method \code{SapiV1SubAccountVirtualSubAccountPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV1SubAccountVirtualSubAccountPostWithHttpInfo(
  sub.account.string,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV2SubAccountFuturesAccountGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV2SubAccountFuturesAccountGet}{}}}
\subsection{Method \code{SapiV2SubAccountFuturesAccountGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV2SubAccountFuturesAccountGet(
  email,
  futures.type,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV2SubAccountFuturesAccountGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV2SubAccountFuturesAccountGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV2SubAccountFuturesAccountGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV2SubAccountFuturesAccountGetWithHttpInfo(
  email,
  futures.type,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV2SubAccountFuturesAccountSummaryGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV2SubAccountFuturesAccountSummaryGet}{}}}
\subsection{Method \code{SapiV2SubAccountFuturesAccountSummaryGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV2SubAccountFuturesAccountSummaryGet(
  futures.type,
  timestamp,
  signature,
  page = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV2SubAccountFuturesAccountSummaryGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV2SubAccountFuturesAccountSummaryGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV2SubAccountFuturesAccountSummaryGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV2SubAccountFuturesAccountSummaryGetWithHttpInfo(
  futures.type,
  timestamp,
  signature,
  page = NULL,
  limit = NULL,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV2SubAccountFuturesPositionRiskGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV2SubAccountFuturesPositionRiskGet}{}}}
\subsection{Method \code{SapiV2SubAccountFuturesPositionRiskGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV2SubAccountFuturesPositionRiskGet(
  email,
  futures.type,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV2SubAccountFuturesPositionRiskGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV2SubAccountFuturesPositionRiskGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV2SubAccountFuturesPositionRiskGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV2SubAccountFuturesPositionRiskGetWithHttpInfo(
  email,
  futures.type,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV3SubAccountAssetsGet"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV3SubAccountAssetsGet}{}}}
\subsection{Method \code{SapiV3SubAccountAssetsGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV3SubAccountAssetsGet(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SapiV3SubAccountAssetsGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SapiV3SubAccountAssetsGetWithHttpInfo}{}}}
\subsection{Method \code{SapiV3SubAccountAssetsGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$SapiV3SubAccountAssetsGetWithHttpInfo(
  email,
  timestamp,
  signature,
  recv.window = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
