% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sub_account_coin_futures_position_risk_delivery_position_risk_vos.R
\docType{class}
\name{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos}
\alias{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos}
\title{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos}
\format{
An \code{R6Class} generator object
}
\description{
SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entryPrice}}{character}

\item{\code{markPrice}}{character}

\item{\code{leverage}}{character}

\item{\code{isolated}}{character}

\item{\code{isolatedWallet}}{character}

\item{\code{isolatedMargin}}{character}

\item{\code{isAutoAddMargin}}{character}

\item{\code{positionSide}}{character}

\item{\code{positionAmount}}{character}

\item{\code{symbol}}{character}

\item{\code{unrealizedProfit}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$new()}}
\item \href{#method-toJSON}{\code{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$toJSON()}}
\item \href{#method-fromJSON}{\code{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$fromJSON()}}
\item \href{#method-toJSONString}{\code{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$fromJSONString()}}
\item \href{#method-clone}{\code{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$new(
  entryPrice,
  markPrice,
  leverage,
  isolated,
  isolatedWallet,
  isolatedMargin,
  isAutoAddMargin,
  positionSide,
  positionAmount,
  symbol,
  unrealizedProfit,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$fromJSON(
  SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVosJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$fromJSONString(
  SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVosJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesPositionRiskDeliveryPositionRiskVos$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
