% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sub_account_coin_futures_summary_delivery_account_summary_resp.R
\docType{class}
\name{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp}
\alias{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp}
\title{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp}
\format{
An \code{R6Class} generator object
}
\description{
SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalMarginBalanceOfBTC}}{character}

\item{\code{totalUnrealizedProfitOfBTC}}{character}

\item{\code{totalWalletBalanceOfBTC}}{character}

\item{\code{asset}}{character}

\item{\code{subAccountList}}{list( \link{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$new()}}
\item \href{#method-toJSON}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$toJSON()}}
\item \href{#method-fromJSON}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$fromJSON()}}
\item \href{#method-toJSONString}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$fromJSONString()}}
\item \href{#method-clone}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$new(
  totalMarginBalanceOfBTC,
  totalUnrealizedProfitOfBTC,
  totalWalletBalanceOfBTC,
  asset,
  subAccountList,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$fromJSON(
  SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$fromJSONString(
  SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryResp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
