% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sub_account_coin_futures_summary_delivery_account_summary_resp_sub_account_list.R
\docType{class}
\name{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList}
\alias{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList}
\title{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList}
\format{
An \code{R6Class} generator object
}
\description{
SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{character}

\item{\code{totalMarginBalance}}{character}

\item{\code{totalUnrealizedProfit}}{character}

\item{\code{totalWalletBalance}}{character}

\item{\code{asset}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$new()}}
\item \href{#method-toJSON}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$toJSON()}}
\item \href{#method-fromJSON}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$fromJSON()}}
\item \href{#method-toJSONString}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$fromJSONString()}}
\item \href{#method-clone}{\code{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$new(
  email,
  totalMarginBalance,
  totalUnrealizedProfit,
  totalWalletBalance,
  asset,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$fromJSON(
  SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountListJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$toJSONString(
  
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$fromJSONString(
  SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountListJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountCOINFuturesSummaryDeliveryAccountSummaryRespSubAccountList$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
