% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_account_usdt_futures_details.R
\docType{class}
\name{SubAccountUSDTFuturesDetails}
\alias{SubAccountUSDTFuturesDetails}
\title{SubAccountUSDTFuturesDetails}
\format{
An \code{R6Class} generator object
}
\description{
SubAccountUSDTFuturesDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{futureAccountResp}}{\link{SubAccountUSDTFuturesDetailsFutureAccountResp}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubAccountUSDTFuturesDetails$new()}}
\item \href{#method-toJSON}{\code{SubAccountUSDTFuturesDetails$toJSON()}}
\item \href{#method-fromJSON}{\code{SubAccountUSDTFuturesDetails$fromJSON()}}
\item \href{#method-toJSONString}{\code{SubAccountUSDTFuturesDetails$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SubAccountUSDTFuturesDetails$fromJSONString()}}
\item \href{#method-clone}{\code{SubAccountUSDTFuturesDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetails$new(futureAccountResp, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetails$fromJSON(SubAccountUSDTFuturesDetailsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetails$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetails$fromJSONString(SubAccountUSDTFuturesDetailsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
