% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sub_account_usdt_futures_details_future_account_resp.R
\docType{class}
\name{SubAccountUSDTFuturesDetailsFutureAccountResp}
\alias{SubAccountUSDTFuturesDetailsFutureAccountResp}
\title{SubAccountUSDTFuturesDetailsFutureAccountResp}
\format{
An \code{R6Class} generator object
}
\description{
SubAccountUSDTFuturesDetailsFutureAccountResp Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{character}

\item{\code{assets}}{list( \link{InlineResponse20063Assets} )}

\item{\code{canDeposit}}{character}

\item{\code{canTrade}}{character}

\item{\code{canWithdraw}}{character}

\item{\code{feeTier}}{integer}

\item{\code{maxWithdrawAmount}}{character}

\item{\code{totalInitialMargin}}{character}

\item{\code{totalMaintenanceMargin}}{character}

\item{\code{totalMarginBalance}}{character}

\item{\code{totalOpenOrderInitialMargin}}{character}

\item{\code{totalPositionInitialMargin}}{character}

\item{\code{totalUnrealizedProfit}}{character}

\item{\code{totalWalletBalance}}{character}

\item{\code{updateTime}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubAccountUSDTFuturesDetailsFutureAccountResp$new()}}
\item \href{#method-toJSON}{\code{SubAccountUSDTFuturesDetailsFutureAccountResp$toJSON()}}
\item \href{#method-fromJSON}{\code{SubAccountUSDTFuturesDetailsFutureAccountResp$fromJSON()}}
\item \href{#method-toJSONString}{\code{SubAccountUSDTFuturesDetailsFutureAccountResp$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SubAccountUSDTFuturesDetailsFutureAccountResp$fromJSONString()}}
\item \href{#method-clone}{\code{SubAccountUSDTFuturesDetailsFutureAccountResp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetailsFutureAccountResp$new(
  email,
  assets,
  canDeposit,
  canTrade,
  canWithdraw,
  feeTier,
  maxWithdrawAmount,
  totalInitialMargin,
  totalMaintenanceMargin,
  totalMarginBalance,
  totalOpenOrderInitialMargin,
  totalPositionInitialMargin,
  totalUnrealizedProfit,
  totalWalletBalance,
  updateTime,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetailsFutureAccountResp$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetailsFutureAccountResp$fromJSON(
  SubAccountUSDTFuturesDetailsFutureAccountRespJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetailsFutureAccountResp$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetailsFutureAccountResp$fromJSONString(
  SubAccountUSDTFuturesDetailsFutureAccountRespJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesDetailsFutureAccountResp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
