% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_account_usdt_futures_position_risk.R
\docType{class}
\name{SubAccountUSDTFuturesPositionRisk}
\alias{SubAccountUSDTFuturesPositionRisk}
\title{SubAccountUSDTFuturesPositionRisk}
\format{
An \code{R6Class} generator object
}
\description{
SubAccountUSDTFuturesPositionRisk Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{futurePositionRiskVos}}{list( \link{InlineResponse20062} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubAccountUSDTFuturesPositionRisk$new()}}
\item \href{#method-toJSON}{\code{SubAccountUSDTFuturesPositionRisk$toJSON()}}
\item \href{#method-fromJSON}{\code{SubAccountUSDTFuturesPositionRisk$fromJSON()}}
\item \href{#method-toJSONString}{\code{SubAccountUSDTFuturesPositionRisk$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SubAccountUSDTFuturesPositionRisk$fromJSONString()}}
\item \href{#method-clone}{\code{SubAccountUSDTFuturesPositionRisk$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesPositionRisk$new(futurePositionRiskVos, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesPositionRisk$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesPositionRisk$fromJSON(
  SubAccountUSDTFuturesPositionRiskJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesPositionRisk$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesPositionRisk$fromJSONString(
  SubAccountUSDTFuturesPositionRiskJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesPositionRisk$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
