% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sub_account_usdt_futures_summary_future_account_summary_resp.R
\docType{class}
\name{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp}
\alias{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp}
\title{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp}
\format{
An \code{R6Class} generator object
}
\description{
SubAccountUSDTFuturesSummaryFutureAccountSummaryResp Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalInitialMargin}}{character}

\item{\code{totalMaintenanceMargin}}{character}

\item{\code{totalMarginBalance}}{character}

\item{\code{totalOpenOrderInitialMargin}}{character}

\item{\code{totalPositionInitialMargin}}{character}

\item{\code{totalUnrealizedProfit}}{character}

\item{\code{totalWalletBalance}}{character}

\item{\code{asset}}{character}

\item{\code{subAccountList}}{list( \link{SubAccountUSDTFuturesSummaryFutureAccountSummaryRespSubAccountList} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$new()}}
\item \href{#method-toJSON}{\code{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$toJSON()}}
\item \href{#method-fromJSON}{\code{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$fromJSON()}}
\item \href{#method-toJSONString}{\code{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$fromJSONString()}}
\item \href{#method-clone}{\code{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$new(
  totalInitialMargin,
  totalMaintenanceMargin,
  totalMarginBalance,
  totalOpenOrderInitialMargin,
  totalPositionInitialMargin,
  totalUnrealizedProfit,
  totalWalletBalance,
  asset,
  subAccountList,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$fromJSON(
  SubAccountUSDTFuturesSummaryFutureAccountSummaryRespJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$fromJSONString(
  SubAccountUSDTFuturesSummaryFutureAccountSummaryRespJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubAccountUSDTFuturesSummaryFutureAccountSummaryResp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
