% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\docType{class}
\name{Ticker}
\alias{Ticker}
\title{Ticker}
\format{
An \code{R6Class} generator object
}
\description{
Ticker Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{symbol}}{character}

\item{\code{priceChange}}{character}

\item{\code{priceChangePercent}}{character}

\item{\code{prevClosePrice}}{character}

\item{\code{lastPrice}}{character}

\item{\code{bidPrice}}{character}

\item{\code{bidQty}}{character}

\item{\code{askPrice}}{character}

\item{\code{askQty}}{character}

\item{\code{openPrice}}{character}

\item{\code{highPrice}}{character}

\item{\code{lowPrice}}{character}

\item{\code{volume}}{character}

\item{\code{quoteVolume}}{character}

\item{\code{openTime}}{integer}

\item{\code{closeTime}}{integer}

\item{\code{firstId}}{integer}

\item{\code{lastId}}{integer}

\item{\code{count}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ticker$new()}}
\item \href{#method-toJSON}{\code{Ticker$toJSON()}}
\item \href{#method-fromJSON}{\code{Ticker$fromJSON()}}
\item \href{#method-toJSONString}{\code{Ticker$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Ticker$fromJSONString()}}
\item \href{#method-clone}{\code{Ticker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$new(
  symbol,
  priceChange,
  priceChangePercent,
  prevClosePrice,
  lastPrice,
  bidPrice,
  bidQty,
  askPrice,
  askQty,
  openPrice,
  highPrice,
  lowPrice,
  volume,
  quoteVolume,
  openTime,
  closeTime,
  firstId,
  lastId,
  count,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$fromJSON(TickerJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$fromJSONString(TickerJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
