% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_stream.R
\name{WebSocketStream}
\alias{WebSocketStream}
\title{R6 class for a websocket market stream}
\description{
R6 class for a websocket market stream

R6 class for a websocket market stream
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WebSocketStream$new()}}
\item \href{#method-Subscribe}{\code{WebSocketStream$Subscribe()}}
\item \href{#method-Unsubscribe}{\code{WebSocketStream$Unsubscribe()}}
\item \href{#method-ListSubscriptions}{\code{WebSocketStream$ListSubscriptions()}}
\item \href{#method-clone}{\code{WebSocketStream$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketStream$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subscribe"></a>}}
\if{latex}{\out{\hypertarget{method-Subscribe}{}}}
\subsection{Method \code{Subscribe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketStream$Subscribe(streams, symbols)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unsubscribe"></a>}}
\if{latex}{\out{\hypertarget{method-Unsubscribe}{}}}
\subsection{Method \code{Unsubscribe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketStream$Unsubscribe(streams, symbols)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListSubscriptions"></a>}}
\if{latex}{\out{\hypertarget{method-ListSubscriptions}{}}}
\subsection{Method \code{ListSubscriptions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketStream$ListSubscriptions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketStream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
