% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response20097.R
\docType{class}
\name{InlineResponse20097}
\alias{InlineResponse20097}
\title{InlineResponse20097}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse20097 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tokenName}}{character}

\item{\code{description}}{character}

\item{\code{underlying}}{character}

\item{\code{tokenIssued}}{character}

\item{\code{basket}}{character}

\item{\code{currentBaskets}}{list( \link{SapiV1BlvtTokenInfoCurrentBaskets} )}

\item{\code{nav}}{character}

\item{\code{realLeverage}}{character}

\item{\code{fundingRate}}{character}

\item{\code{dailyManagementFee}}{character}

\item{\code{purchaseFeePct}}{character}

\item{\code{dailyPurchaseLimit}}{character}

\item{\code{redeemFeePct}}{character}

\item{\code{dailyRedeemLimit}}{character}

\item{\code{timestamp}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20097$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20097$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20097$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20097$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20097$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20097$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097$new(
  tokenName,
  description,
  underlying,
  tokenIssued,
  basket,
  currentBaskets,
  nav,
  realLeverage,
  fundingRate,
  dailyManagementFee,
  purchaseFeePct,
  dailyPurchaseLimit,
  redeemFeePct,
  dailyRedeemLimit,
  timestamp,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097$fromJSON(InlineResponse20097Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097$fromJSONString(InlineResponse20097Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20097$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
